Separating interface implementations based on clientsets
