+++ k8s/k8s.go
@@ -4,36 +4,25 @@ import (
 	"bytes"
 	"fmt"
 	"log"
-	"net"
-	"os"
 	"strings"
 	"sync"
 	"time"
 
 	prometheusclient "github.com/coreos/prometheus-operator/pkg/client/versioned"
 	snap_v1 "github.com/kubernetes-incubator/external-storage/snapshot/pkg/apis/crd/v1"
-	snap_client "github.com/kubernetes-incubator/external-storage/snapshot/pkg/client"
-	aut_v1alpaha1 "github.com/libopenstorage/autopilot/pkg/apis/autopilot/v1alpha1"
 	autopilotclientset "github.com/libopenstorage/autopilot/pkg/client/clientset/versioned"
-	"github.com/libopenstorage/stork/pkg/apis/stork/v1alpha1"
 	storkclientset "github.com/libopenstorage/stork/pkg/client/clientset/versioned"
-	ocp_appsv1_api "github.com/openshift/api/apps/v1"
-	ocp_securityv1_api "github.com/openshift/api/security/v1"
 	ocp_clientset "github.com/openshift/client-go/apps/clientset/versioned"
-	ocp_appsv1_client "github.com/openshift/client-go/apps/clientset/versioned/typed/apps/v1"
 	ocp_security_clientset "github.com/openshift/client-go/security/clientset/versioned"
-	ocp_securityv1_client "github.com/openshift/client-go/security/clientset/versioned/typed/security/v1"
 	"github.com/portworx/sched-ops/task"
-	talisman_v1beta2 "github.com/portworx/talisman/pkg/apis/portworx/v1beta2"
 	talismanclientset "github.com/portworx/talisman/pkg/client/clientset/versioned"
 	"github.com/sirupsen/logrus"
-	apps_api "k8s.io/api/apps/v1"
+	appsv1 "k8s.io/api/apps/v1"
 	batch_v1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
 	rbac_v1 "k8s.io/api/rbac/v1"
-	storage_api "k8s.io/api/storage/v1"
+	storagev1 "k8s.io/api/storage/v1"
 	storagev1beta1 "k8s.io/api/storage/v1beta1"
-	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -42,16 +31,13 @@ import (
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
-	appsv1_client "k8s.io/client-go/kubernetes/typed/apps/v1"
 	_ "k8s.io/client-go/plugin/pkg/client/auth" // needed to initialize plugins on cloud providers
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/tools/remotecommand"
 )
 
@@ -64,11 +50,6 @@ const (
 )
 
 var deleteForegroundPolicy = meta_v1.DeletePropagationForeground
-
-var (
-	// ErrPodsNotFound error returned when pod or pods could not be found
-	ErrPodsNotFound = fmt.Errorf("Pod(s) not found")
-)
 
 // Ops is an interface to perform any kubernetes related operations
 type Ops interface {
@@ -86,6 +67,7 @@ type Ops interface {
 	PersistentVolumeClaimOps
 	VolumeAttachmentOps
 	SnapshotOps
+	SnapshotScheduleOps
 	GroupSnapshotOps
 	RuleOps
 	SecretOps
@@ -109,16 +91,6 @@ type Ops interface {
 	GetVersion() (*version.Info, error)
 	// private methods for unit tests
 	privateMethods
-}
-
-// SecurityContextConstraintsOps is an interface to list, get and update security context constraints
-type SecurityContextConstraintsOps interface {
-	// ListSecurityContextConstraints returns the list of all SecurityContextConstraints, and an error if there is any.
-	ListSecurityContextConstraints() (*ocp_securityv1_api.SecurityContextConstraintsList, error)
-	// GetSecurityContextConstraints takes name of the securityContextConstraints and returns the corresponding securityContextConstraints object, and an error if there is any.
-	GetSecurityContextConstraints(string) (*ocp_securityv1_api.SecurityContextConstraints, error)
-	// UpdateSecurityContextConstraints takes the representation of a securityContextConstraints and updates it. Returns the server's representation of the securityContextConstraints, and an error, if there is any.
-	UpdateSecurityContextConstraints(*ocp_securityv1_api.SecurityContextConstraints) (*ocp_securityv1_api.SecurityContextConstraints, error)
 }
 
 // EventOps is an interface to put and get k8s events
@@ -195,93 +167,69 @@ type ServiceOps interface {
 // StatefulSetOps is an interface to perform k8s stateful set operations
 type StatefulSetOps interface {
 	// ListStatefulSets lists all the statefulsets for a given namespace
-	ListStatefulSets(namespace string) (*apps_api.StatefulSetList, error)
+	ListStatefulSets(namespace string) (*appsv1.StatefulSetList, error)
 	// GetStatefulSet returns a statefulset for given name and namespace
-	GetStatefulSet(name, namespace string) (*apps_api.StatefulSet, error)
+	GetStatefulSet(name, namespace string) (*appsv1.StatefulSet, error)
 	// CreateStatefulSet creates the given statefulset
-	CreateStatefulSet(ss *apps_api.StatefulSet) (*apps_api.StatefulSet, error)
+	CreateStatefulSet(ss *appsv1.StatefulSet) (*appsv1.StatefulSet, error)
 	// UpdateStatefulSet creates the given statefulset
-	UpdateStatefulSet(ss *apps_api.StatefulSet) (*apps_api.StatefulSet, error)
+	UpdateStatefulSet(ss *appsv1.StatefulSet) (*appsv1.StatefulSet, error)
 	// DeleteStatefulSet deletes the given statefulset
 	DeleteStatefulSet(name, namespace string) error
 	// ValidateStatefulSet validates the given statefulset if it's running and healthy within the given timeout
-	ValidateStatefulSet(ss *apps_api.StatefulSet, timeout time.Duration) error
+	ValidateStatefulSet(ss *appsv1.StatefulSet, timeout time.Duration) error
 	// ValidateTerminatedStatefulSet validates if given deployment is terminated
-	ValidateTerminatedStatefulSet(ss *apps_api.StatefulSet, timeout, retryInterval time.Duration) error
+	ValidateTerminatedStatefulSet(ss *appsv1.StatefulSet, timeout, retryInterval time.Duration) error
 	// GetStatefulSetPods returns pods for the given statefulset
-	GetStatefulSetPods(ss *apps_api.StatefulSet) ([]v1.Pod, error)
+	GetStatefulSetPods(ss *appsv1.StatefulSet) ([]v1.Pod, error)
 	// DescribeStatefulSet gets status of the statefulset
-	DescribeStatefulSet(name, namespace string) (*apps_api.StatefulSetStatus, error)
+	DescribeStatefulSet(name, namespace string) (*appsv1.StatefulSetStatus, error)
 	// GetStatefulSetsUsingStorageClass returns all statefulsets using given storage class
-	GetStatefulSetsUsingStorageClass(scName string) ([]apps_api.StatefulSet, error)
+	GetStatefulSetsUsingStorageClass(scName string) ([]appsv1.StatefulSet, error)
 	// GetPVCsForStatefulSet returns all the PVCs for given stateful set
-	GetPVCsForStatefulSet(ss *apps_api.StatefulSet) (*v1.PersistentVolumeClaimList, error)
+	GetPVCsForStatefulSet(ss *appsv1.StatefulSet) (*v1.PersistentVolumeClaimList, error)
 	// ValidatePVCsForStatefulSet validates the PVCs for the given stateful set
-	ValidatePVCsForStatefulSet(ss *apps_api.StatefulSet, timeout, retryInterval time.Duration) error
+	ValidatePVCsForStatefulSet(ss *appsv1.StatefulSet, timeout, retryInterval time.Duration) error
 }
 
 // DeploymentOps is an interface to perform k8s deployment operations
 type DeploymentOps interface {
 	// ListDeployments lists all deployments for the given namespace
-	ListDeployments(namespace string, options meta_v1.ListOptions) (*apps_api.DeploymentList, error)
+	ListDeployments(namespace string, options meta_v1.ListOptions) (*appsv1.DeploymentList, error)
 	// GetDeployment returns a deployment for the give name and namespace
-	GetDeployment(name, namespace string) (*apps_api.Deployment, error)
+	GetDeployment(name, namespace string) (*appsv1.Deployment, error)
 	// CreateDeployment creates the given deployment
-	CreateDeployment(*apps_api.Deployment) (*apps_api.Deployment, error)
+	CreateDeployment(*appsv1.Deployment) (*appsv1.Deployment, error)
 	// UpdateDeployment updates the given deployment
-	UpdateDeployment(*apps_api.Deployment) (*apps_api.Deployment, error)
+	UpdateDeployment(*appsv1.Deployment) (*appsv1.Deployment, error)
 	// DeleteDeployment deletes the given deployment
 	DeleteDeployment(name, namespace string) error
 	// ValidateDeployment validates the given deployment if it's running and healthy
-	ValidateDeployment(deployment *apps_api.Deployment, timeout, retryInterval time.Duration) error
+	ValidateDeployment(deployment *appsv1.Deployment, timeout, retryInterval time.Duration) error
 	// ValidateTerminatedDeployment validates if given deployment is terminated
-	ValidateTerminatedDeployment(*apps_api.Deployment, time.Duration, time.Duration) error
+	ValidateTerminatedDeployment(*appsv1.Deployment, time.Duration, time.Duration) error
 	// GetDeploymentPods returns pods for the given deployment
-	GetDeploymentPods(*apps_api.Deployment) ([]v1.Pod, error)
+	GetDeploymentPods(*appsv1.Deployment) ([]v1.Pod, error)
 	// DescribeDeployment gets the deployment status
-	DescribeDeployment(name, namespace string) (*apps_api.DeploymentStatus, error)
+	DescribeDeployment(name, namespace string) (*appsv1.DeploymentStatus, error)
 	// GetDeploymentsUsingStorageClass returns all deployments using the given storage class
-	GetDeploymentsUsingStorageClass(scName string) ([]apps_api.Deployment, error)
-}
-
-// DeploymentConfigOps is an interface to perform ocp deployment config operations
-type DeploymentConfigOps interface {
-	// ListDeploymentConfigs lists all deployments for the given namespace
-	ListDeploymentConfigs(namespace string) (*ocp_appsv1_api.DeploymentConfigList, error)
-	// GetDeploymentConfig returns a deployment for the give name and namespace
-	GetDeploymentConfig(name, namespace string) (*ocp_appsv1_api.DeploymentConfig, error)
-	// CreateDeploymentConfig creates the given deployment
-	CreateDeploymentConfig(*ocp_appsv1_api.DeploymentConfig) (*ocp_appsv1_api.DeploymentConfig, error)
-	// UpdateDeploymentConfig updates the given deployment
-	UpdateDeploymentConfig(*ocp_appsv1_api.DeploymentConfig) (*ocp_appsv1_api.DeploymentConfig, error)
-	// DeleteDeploymentConfig deletes the given deployment
-	DeleteDeploymentConfig(name, namespace string) error
-	// ValidateDeploymentConfig validates the given deployment if it's running and healthy
-	ValidateDeploymentConfig(deployment *ocp_appsv1_api.DeploymentConfig, timeout, retryInterval time.Duration) error
-	// ValidateTerminatedDeploymentConfig validates if given deployment is terminated
-	ValidateTerminatedDeploymentConfig(*ocp_appsv1_api.DeploymentConfig) error
-	// GetDeploymentConfigPods returns pods for the given deployment
-	GetDeploymentConfigPods(*ocp_appsv1_api.DeploymentConfig) ([]v1.Pod, error)
-	// DescribeDeploymentConfig gets the deployment status
-	DescribeDeploymentConfig(name, namespace string) (*ocp_appsv1_api.DeploymentConfigStatus, error)
-	// GetDeploymentConfigsUsingStorageClass returns all deployments using the given storage class
-	GetDeploymentConfigsUsingStorageClass(scName string) ([]ocp_appsv1_api.DeploymentConfig, error)
+	GetDeploymentsUsingStorageClass(scName string) ([]appsv1.Deployment, error)
 }
 
 // DaemonSetOps is an interface to perform k8s daemon set operations
 type DaemonSetOps interface {
 	// CreateDaemonSet creates the given daemonset
-	CreateDaemonSet(ds *apps_api.DaemonSet) (*apps_api.DaemonSet, error)
+	CreateDaemonSet(ds *appsv1.DaemonSet) (*appsv1.DaemonSet, error)
 	// ListDaemonSets lists all daemonsets in given namespace
-	ListDaemonSets(namespace string, listOpts meta_v1.ListOptions) ([]apps_api.DaemonSet, error)
+	ListDaemonSets(namespace string, listOpts meta_v1.ListOptions) ([]appsv1.DaemonSet, error)
 	// GetDaemonSet gets the the daemon set with given name
-	GetDaemonSet(string, string) (*apps_api.DaemonSet, error)
+	GetDaemonSet(string, string) (*appsv1.DaemonSet, error)
 	// ValidateDaemonSet checks if the given daemonset is ready within given timeout
 	ValidateDaemonSet(name, namespace string, timeout time.Duration) error
 	// GetDaemonSetPods returns list of pods for the daemonset
-	GetDaemonSetPods(*apps_api.DaemonSet) ([]v1.Pod, error)
+	GetDaemonSetPods(*appsv1.DaemonSet) ([]v1.Pod, error)
 	// UpdateDaemonSet updates the given daemon set and returns the updated ds
-	UpdateDaemonSet(*apps_api.DaemonSet) (*apps_api.DaemonSet, error)
+	UpdateDaemonSet(*appsv1.DaemonSet) (*appsv1.DaemonSet, error)
 	// DeleteDaemonSet deletes the given daemonset
 	DeleteDaemonSet(name, namespace string) error
 }
@@ -389,19 +337,19 @@ type PodOps interface {
 // StorageClassOps is an interface to perform k8s storage class operations
 type StorageClassOps interface {
 	// GetStorageClasses returns all storageClasses that match given optional label selector
-	GetStorageClasses(labelSelector map[string]string) (*storage_api.StorageClassList, error)
+	GetStorageClasses(labelSelector map[string]string) (*storagev1.StorageClassList, error)
 	// GetStorageClass returns the storage class for the give namme
-	GetStorageClass(name string) (*storage_api.StorageClass, error)
+	GetStorageClass(name string) (*storagev1.StorageClass, error)
 	// CreateStorageClass creates the given storage class
-	CreateStorageClass(sc *storage_api.StorageClass) (*storage_api.StorageClass, error)
+	CreateStorageClass(sc *storagev1.StorageClass) (*storagev1.StorageClass, error)
 	// DeleteStorageClass deletes the given storage class
 	DeleteStorageClass(name string) error
 	// GetStorageClassParams returns the parameters of the given sc in the native map format
-	GetStorageClassParams(sc *storage_api.StorageClass) (map[string]string, error)
+	GetStorageClassParams(sc *storagev1.StorageClass) (map[string]string, error)
 	// ValidateStorageClass validates the given storage class
 	// TODO: This is currently the same as GetStorageClass. If no one is using it,
 	// we should remove this method
-	ValidateStorageClass(name string) (*storage_api.StorageClass, error)
+	ValidateStorageClass(name string) (*storagev1.StorageClass, error)
 }
 
 // PersistentVolumeClaimOps is an interface to perform k8s PVC operations
@@ -466,22 +414,6 @@ type SnapshotOps interface {
 	DeleteSnapshotData(name string) error
 	// ValidateSnapshotData validates the given snapshot data object
 	ValidateSnapshotData(name string, retry bool, timeout, retryInterval time.Duration) error
-	// GetSnapshotSchedule gets the SnapshotSchedule
-	GetSnapshotSchedule(string, string) (*v1alpha1.VolumeSnapshotSchedule, error)
-	// CreateSnapshotSchedule creates a SnapshotSchedule
-	CreateSnapshotSchedule(*v1alpha1.VolumeSnapshotSchedule) (*v1alpha1.VolumeSnapshotSchedule, error)
-	// UpdateSnapshotSchedule updates the SnapshotSchedule
-	UpdateSnapshotSchedule(*v1alpha1.VolumeSnapshotSchedule) (*v1alpha1.VolumeSnapshotSchedule, error)
-	// ListSnapshotSchedules lists all the SnapshotSchedules
-	ListSnapshotSchedules(string) (*v1alpha1.VolumeSnapshotScheduleList, error)
-	// DeleteSnapshotSchedule deletes the SnapshotSchedule
-	DeleteSnapshotSchedule(string, string) error
-	// ValidateSnapshotSchedule validates the given SnapshotSchedule. It checks the status of each of
-	// the snapshots triggered for this schedule and returns a map of successfull snapshots. The key of the
-	// map will be the schedule type and value will be list of snapshots for that schedule type.
-	// The caller is expected to validate if the returned map has all snapshots expected at that point of time
-	ValidateSnapshotSchedule(string, string, time.Duration, time.Duration) (
-		map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledVolumeSnapshotStatus, error)
 }
 
 // VolumeAttachmentOps is an interface to perform k8s VolumeAttachmentOps operations
@@ -496,52 +428,6 @@ type VolumeAttachmentOps interface {
 	UpdateVolumeAttachment(*storagev1beta1.VolumeAttachment) (*storagev1beta1.VolumeAttachment, error)
 	// UpdateVolumeAttachmentStatus updates a volume attachment status
 	UpdateVolumeAttachmentStatus(*storagev1beta1.VolumeAttachment) (*storagev1beta1.VolumeAttachment, error)
-}
-
-// GroupSnapshotOps is an interface to perform k8s GroupVolumeSnapshot operations
-type GroupSnapshotOps interface {
-	// GetGroupSnapshot returns the group snapshot for the given name and namespace
-	GetGroupSnapshot(name, namespace string) (*v1alpha1.GroupVolumeSnapshot, error)
-	// ListGroupSnapshots lists all group snapshots for the given namespace
-	ListGroupSnapshots(namespace string) (*v1alpha1.GroupVolumeSnapshotList, error)
-	// CreateGroupSnapshot creates the given group snapshot
-	CreateGroupSnapshot(*v1alpha1.GroupVolumeSnapshot) (*v1alpha1.GroupVolumeSnapshot, error)
-	// UpdateGroupSnapshot updates the given group snapshot
-	UpdateGroupSnapshot(*v1alpha1.GroupVolumeSnapshot) (*v1alpha1.GroupVolumeSnapshot, error)
-	// DeleteGroupSnapshot deletes the group snapshot with the given name and namespace
-	DeleteGroupSnapshot(name, namespace string) error
-	// ValidateGroupSnapshot checks if the group snapshot with given name and namespace is in ready state
-	//  If retry is true, the validation will be retried with given timeout and retry internal
-	ValidateGroupSnapshot(name, namespace string, retry bool, timeout, retryInterval time.Duration) error
-	// GetSnapshotsForGroupSnapshot returns all child snapshots for the group snapshot
-	GetSnapshotsForGroupSnapshot(name, namespace string) ([]*snap_v1.VolumeSnapshot, error)
-}
-
-// VolumeSnapshotRestoreOps is interface to perform isnapshot restore using CRD
-type VolumeSnapshotRestoreOps interface {
-	// CreateVolumeSnapshotRestore restore snapshot to pvc specifed in CRD, if no pvcs defined we restore to
-	// parent volumes
-	CreateVolumeSnapshotRestore(snap *v1alpha1.VolumeSnapshotRestore) (*v1alpha1.VolumeSnapshotRestore, error)
-	// UpdateVolumeSnapshotRestore updates given volumesnapshorestore CRD
-	UpdateVolumeSnapshotRestore(snap *v1alpha1.VolumeSnapshotRestore) (*v1alpha1.VolumeSnapshotRestore, error)
-	// GetVolumeSnapshotRestore returns details of given restore crd status
-	GetVolumeSnapshotRestore(name, namespace string) (*v1alpha1.VolumeSnapshotRestore, error)
-	// ListVolumeSnapshotRestore return list of volumesnapshotrestores in given namespaces
-	ListVolumeSnapshotRestore(namespace string) (*v1alpha1.VolumeSnapshotRestoreList, error)
-	// DeleteVolumeSnapshotRestore delete given volumesnapshotrestore CRD
-	DeleteVolumeSnapshotRestore(name, namespace string) error
-	// ValidateVolumeSnapshotRestore validates given volumesnapshotrestore CRD
-	ValidateVolumeSnapshotRestore(name, namespace string, timeout, retry time.Duration) error
-}
-
-// RuleOps is an interface to perform operations for k8s stork rule
-type RuleOps interface {
-	// GetRule fetches the given stork rule
-	GetRule(name, namespace string) (*v1alpha1.Rule, error)
-	// CreateRule creates the given stork rule
-	CreateRule(rule *v1alpha1.Rule) (*v1alpha1.Rule, error)
-	// DeleteRule deletes the given stork rule
-	DeleteRule(name, namespace string) error
 }
 
 // SecretOps is an interface to perform k8s Secret operations
@@ -572,110 +458,6 @@ type ConfigMapOps interface {
 	WatchConfigMap(configMap *v1.ConfigMap, fn WatchFunc) error
 }
 
-// CRDOps is an interface to perfrom k8s Customer Resource operations
-type CRDOps interface {
-	// CreateCRD creates the given custom resource
-	// This API will be deprecated soon. Use RegisterCRD instead
-	CreateCRD(resource CustomResource) error
-	// RegisterCRD creates the given custom resource
-	RegisterCRD(crd *apiextensionsv1beta1.CustomResourceDefinition) error
-	// ValidateCRD checks if the given CRD is registered
-	ValidateCRD(resource CustomResource, timeout, retryInterval time.Duration) error
-	// DeleteCRD deletes the CRD for the given complete name (plural.group)
-	DeleteCRD(fullName string) error
-}
-
-// ClusterPairOps is an interface to perfrom k8s ClusterPair operations
-type ClusterPairOps interface {
-	// CreateClusterPair creates the ClusterPair
-	CreateClusterPair(*v1alpha1.ClusterPair) (*v1alpha1.ClusterPair, error)
-	// GetClusterPair gets the ClusterPair
-	GetClusterPair(string, string) (*v1alpha1.ClusterPair, error)
-	// ListClusterPairs gets all the ClusterPairs
-	ListClusterPairs(string) (*v1alpha1.ClusterPairList, error)
-	// UpdateClusterPair updates the ClusterPair
-	UpdateClusterPair(*v1alpha1.ClusterPair) (*v1alpha1.ClusterPair, error)
-	// DeleteClusterPair deletes the ClusterPair
-	DeleteClusterPair(string, string) error
-	// ValidateClusterPair validates clusterpair status
-	ValidateClusterPair(string, string, time.Duration, time.Duration) error
-}
-
-// ClusterDomainsOps is an interface to perform k8s ClusterDomains operations
-type ClusterDomainsOps interface {
-	// CreateClusterDomainsStatus creates the ClusterDomainStatus
-	CreateClusterDomainsStatus(*v1alpha1.ClusterDomainsStatus) (*v1alpha1.ClusterDomainsStatus, error)
-	// GetClusterDomainsStatus gets the ClusterDomainsStatus
-	GetClusterDomainsStatus(string) (*v1alpha1.ClusterDomainsStatus, error)
-	// UpdateClusterDomainsStatus updates the ClusterDomainsStatus
-	UpdateClusterDomainsStatus(*v1alpha1.ClusterDomainsStatus) (*v1alpha1.ClusterDomainsStatus, error)
-	// DeleteClusterDomainsStatus deletes the ClusterDomainsStatus
-	DeleteClusterDomainsStatus(string) error
-	// ListClusterDomainStatuses lists ClusterDomainsStatus
-	ListClusterDomainStatuses() (*v1alpha1.ClusterDomainsStatusList, error)
-	// ValidateClusterDomainsStatus validates the ClusterDomainsStatus
-	ValidateClusterDomainsStatus(string, map[string]bool, time.Duration, time.Duration) error
-
-	// CreateClusterDomainUpdate creates the ClusterDomainUpdate
-	CreateClusterDomainUpdate(*v1alpha1.ClusterDomainUpdate) (*v1alpha1.ClusterDomainUpdate, error)
-	// GetClusterDomainUpdate gets the ClusterDomainUpdate
-	GetClusterDomainUpdate(string) (*v1alpha1.ClusterDomainUpdate, error)
-	// UpdateClusterDomainUpdate updates the ClusterDomainUpdate
-	UpdateClusterDomainUpdate(*v1alpha1.ClusterDomainUpdate) (*v1alpha1.ClusterDomainUpdate, error)
-	// DeleteClusterDomainUpdate deletes the ClusterDomainUpdate
-	DeleteClusterDomainUpdate(string) error
-	// ValidateClusterDomainUpdate validates ClusterDomainUpdate
-	ValidateClusterDomainUpdate(string, time.Duration, time.Duration) error
-	// ListClusterDomainUpdates lists ClusterDomainUpdates
-	ListClusterDomainUpdates() (*v1alpha1.ClusterDomainUpdateList, error)
-}
-
-// AutopilotRuleOps is an interface to perform k8s AutopilotRule operations
-type AutopilotRuleOps interface {
-	// CreateAutopilotRule creates the AutopilotRule object
-	CreateAutopilotRule(*aut_v1alpaha1.AutopilotRule) (*aut_v1alpaha1.AutopilotRule, error)
-	// GetAutopilotRule gets the AutopilotRule for the provided name
-	GetAutopilotRule(string) (*aut_v1alpaha1.AutopilotRule, error)
-	// UpdateAutopilotRule updates the AutopilotRule
-	UpdateAutopilotRule(*aut_v1alpaha1.AutopilotRule) (*aut_v1alpaha1.AutopilotRule, error)
-	// DeleteAutopilotRule deletes the AutopilotRule of the given name
-	DeleteAutopilotRule(string) error
-	// ListAutopilotRules lists AutopilotRules
-	ListAutopilotRules() (*aut_v1alpaha1.AutopilotRuleList, error)
-}
-
-// MigrationOps is an interface to perfrom k8s Migration operations
-type MigrationOps interface {
-	// CreateMigration creates the Migration
-	CreateMigration(*v1alpha1.Migration) (*v1alpha1.Migration, error)
-	// GetMigration gets the Migration
-	GetMigration(string, string) (*v1alpha1.Migration, error)
-	// ListMigrations lists all the Migrations
-	ListMigrations(string) (*v1alpha1.MigrationList, error)
-	// UpdateMigration updates the Migration
-	UpdateMigration(*v1alpha1.Migration) (*v1alpha1.Migration, error)
-	// DeleteMigration deletes the Migration
-	DeleteMigration(string, string) error
-	// ValidateMigration validate the Migration status
-	ValidateMigration(string, string, time.Duration, time.Duration) error
-	// GetMigrationSchedule gets the MigrationSchedule
-	GetMigrationSchedule(string, string) (*v1alpha1.MigrationSchedule, error)
-	// CreateMigrationSchedule creates a MigrationSchedule
-	CreateMigrationSchedule(*v1alpha1.MigrationSchedule) (*v1alpha1.MigrationSchedule, error)
-	// UpdateMigrationSchedule updates the MigrationSchedule
-	UpdateMigrationSchedule(*v1alpha1.MigrationSchedule) (*v1alpha1.MigrationSchedule, error)
-	// ListMigrationSchedules lists all the MigrationSchedules
-	ListMigrationSchedules(string) (*v1alpha1.MigrationScheduleList, error)
-	// DeleteMigrationSchedule deletes the MigrationSchedule
-	DeleteMigrationSchedule(string, string) error
-	// ValidateMigrationSchedule validates the given MigrationSchedule. It checks the status of each of
-	// the migrations triggered for this schedule and returns a map of successfull migrations. The key of the
-	// map will be the schedule type and value will be list of migrations for that schedule type.
-	// The caller is expected to validate if the returned map has all migrations expected at that point of time
-	ValidateMigrationSchedule(string, string, time.Duration, time.Duration) (
-		map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledMigrationStatus, error)
-}
-
 // ObjectOps is an interface to perform generic Object operations
 type ObjectOps interface {
 	// GetObject returns the latest object given a generic Object
@@ -684,176 +466,8 @@ type ObjectOps interface {
 	UpdateObject(object runtime.Object) (runtime.Object, error)
 }
 
-// SchedulePolicyOps is an interface to manage SchedulePolicy Object
-type SchedulePolicyOps interface {
-	// CreateSchedulePolicy creates a SchedulePolicy
-	CreateSchedulePolicy(*v1alpha1.SchedulePolicy) (*v1alpha1.SchedulePolicy, error)
-	// GetSchedulePolicy gets the SchedulePolicy
-	GetSchedulePolicy(string) (*v1alpha1.SchedulePolicy, error)
-	// ListSchedulePolicies lists all the SchedulePolicies
-	ListSchedulePolicies() (*v1alpha1.SchedulePolicyList, error)
-	// UpdateSchedulePolicy updates the SchedulePolicy
-	UpdateSchedulePolicy(*v1alpha1.SchedulePolicy) (*v1alpha1.SchedulePolicy, error)
-	// DeleteSchedulePolicy deletes the SchedulePolicy
-	DeleteSchedulePolicy(string) error
-}
-
-// VolumePlacementStrategyOps is an interface to perform CRUD volume placememt strategy ops
-type VolumePlacementStrategyOps interface {
-	// CreateVolumePlacementStrategy creates a new volume placement strategy
-	CreateVolumePlacementStrategy(spec *talisman_v1beta2.VolumePlacementStrategy) (*talisman_v1beta2.VolumePlacementStrategy, error)
-	// UpdateVolumePlacementStrategy updates an existing volume placement strategy
-	UpdateVolumePlacementStrategy(spec *talisman_v1beta2.VolumePlacementStrategy) (*talisman_v1beta2.VolumePlacementStrategy, error)
-	// ListVolumePlacementStrategies lists all volume placement strategies
-	ListVolumePlacementStrategies() (*talisman_v1beta2.VolumePlacementStrategyList, error)
-	// DeleteVolumePlacementStrategy deletes the volume placement strategy with given name
-	DeleteVolumePlacementStrategy(name string) error
-	// GetVolumePlacementStrategy returns the volume placememt strategy with given name
-	GetVolumePlacementStrategy(name string) (*talisman_v1beta2.VolumePlacementStrategy, error)
-}
-
-// BackupLocationOps is an interface to perfrom k8s BackupLocation operations
-type BackupLocationOps interface {
-	// CreateBackupLocation creates the BackupLocation
-	CreateBackupLocation(*v1alpha1.BackupLocation) (*v1alpha1.BackupLocation, error)
-	// GetBackupLocation gets the BackupLocation
-	GetBackupLocation(string, string) (*v1alpha1.BackupLocation, error)
-	// ListBackupLocations lists all the BackupLocations
-	ListBackupLocations(string) (*v1alpha1.BackupLocationList, error)
-	// UpdateBackupLocation updates the BackupLocation
-	UpdateBackupLocation(*v1alpha1.BackupLocation) (*v1alpha1.BackupLocation, error)
-	// DeleteBackupLocation deletes the BackupLocation
-	DeleteBackupLocation(string, string) error
-	// ValidateBackupLocation validates the BackupLocation
-	ValidateBackupLocation(string, string, time.Duration, time.Duration) error
-}
-
-// ApplicationBackupRestoreOps is an interface to perfrom k8s Application Backup
-// and Restore operations
-type ApplicationBackupRestoreOps interface {
-	// CreateApplicationBackup creates the ApplicationBackup
-	CreateApplicationBackup(*v1alpha1.ApplicationBackup) (*v1alpha1.ApplicationBackup, error)
-	// GetApplicationBackup gets the ApplicationBackup
-	GetApplicationBackup(string, string) (*v1alpha1.ApplicationBackup, error)
-	// ListApplicationBackups lists all the ApplicationBackups
-	ListApplicationBackups(string) (*v1alpha1.ApplicationBackupList, error)
-	// UpdateApplicationBackup updates the ApplicationBackup
-	UpdateApplicationBackup(*v1alpha1.ApplicationBackup) (*v1alpha1.ApplicationBackup, error)
-	// DeleteApplicationBackup deletes the ApplicationBackup
-	DeleteApplicationBackup(string, string) error
-	// ValidateApplicationBackup validates the ApplicationBackup
-	ValidateApplicationBackup(string, string, time.Duration, time.Duration) error
-	// CreateApplicationRestore creates the ApplicationRestore
-	CreateApplicationRestore(*v1alpha1.ApplicationRestore) (*v1alpha1.ApplicationRestore, error)
-	// GetApplicationRestore gets the ApplicationRestore
-	GetApplicationRestore(string, string) (*v1alpha1.ApplicationRestore, error)
-	// ListApplicationRestores lists all the ApplicationRestores
-	ListApplicationRestores(string) (*v1alpha1.ApplicationRestoreList, error)
-	// UpdateApplicationRestore updates the ApplicationRestore
-	UpdateApplicationRestore(*v1alpha1.ApplicationRestore) (*v1alpha1.ApplicationRestore, error)
-	// DeleteApplicationRestore deletes the ApplicationRestore
-	DeleteApplicationRestore(string, string) error
-	// ValidateApplicationRestore validates the ApplicationRestore
-	ValidateApplicationRestore(string, string, time.Duration, time.Duration) error
-	// GetApplicationBackupSchedule gets the ApplicationBackupSchedule
-	GetApplicationBackupSchedule(string, string) (*v1alpha1.ApplicationBackupSchedule, error)
-	// CreateApplicationBackupSchedule creates an ApplicationBackupSchedule
-	CreateApplicationBackupSchedule(*v1alpha1.ApplicationBackupSchedule) (*v1alpha1.ApplicationBackupSchedule, error)
-	// UpdateApplicationBackupSchedule updates the ApplicationBackupSchedule
-	UpdateApplicationBackupSchedule(*v1alpha1.ApplicationBackupSchedule) (*v1alpha1.ApplicationBackupSchedule, error)
-	// ListApplicationBackupSchedules lists all the ApplicationBackupSchedules
-	ListApplicationBackupSchedules(string) (*v1alpha1.ApplicationBackupScheduleList, error)
-	// DeleteApplicationBackupSchedule deletes the ApplicationBackupSchedule
-	DeleteApplicationBackupSchedule(string, string) error
-	// ValidateApplicationBackupSchedule validates the given ApplicationBackupSchedule. It checks the status of each of
-	// the backups triggered for this schedule and returns a map of successfull backups. The key of the
-	// map will be the schedule type and value will be list of backups for that schedule type.
-	// The caller is expected to validate if the returned map has all backups expected at that point of time
-	ValidateApplicationBackupSchedule(string, string, time.Duration, time.Duration) (
-		map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledApplicationBackupStatus, error)
-}
-
-// ApplicationCloneOps is an interface to perform k8s Application Clone operations
-type ApplicationCloneOps interface {
-	// CreateApplicationClone creates the ApplicationClone
-	CreateApplicationClone(*v1alpha1.ApplicationClone) (*v1alpha1.ApplicationClone, error)
-	// GetApplicationClone gets the ApplicationClone
-	GetApplicationClone(string, string) (*v1alpha1.ApplicationClone, error)
-	// ListApplicationClones lists all the ApplicationClones
-	ListApplicationClones(string) (*v1alpha1.ApplicationCloneList, error)
-	// UpdateApplicationClone updates the ApplicationClone
-	UpdateApplicationClone(*v1alpha1.ApplicationClone) (*v1alpha1.ApplicationClone, error)
-	// DeleteApplicationClone deletes the ApplicationClone
-	DeleteApplicationClone(string, string) error
-	// ValidateApplicationClone validates the ApplicationClone
-	ValidateApplicationClone(string, string, time.Duration, time.Duration) error
-}
-
-// ClientSetter is an interface to allow setting different clients on the Ops object
-type ClientSetter interface {
-	// SetConfig sets the config and resets the client
-	SetConfig(config *rest.Config)
-	// SetConfigFromPath sets the config from a kubeconfig file
-	SetConfigFromPath(configPath string) error
-	// SetClient set the k8s clients
-	SetClient(
-		kubernetes.Interface,
-		rest.Interface,
-		storkclientset.Interface,
-		apiextensionsclient.Interface,
-		dynamic.Interface,
-		ocp_clientset.Interface,
-		ocp_security_clientset.Interface,
-		autopilotclientset.Interface,
-	)
-	// SetBaseClient sets the kubernetes clientset
-	SetBaseClient(kubernetes.Interface)
-	// SetSnapshotClient sets the snapshot clientset
-	SetSnapshotClient(rest.Interface)
-	// SetStorkClient sets the stork clientset
-	SetStorkClient(storkclientset.Interface)
-	// SetApiExtensionsClient sets the api extensions clientset
-	SetApiExtensionsClient(apiextensionsclient.Interface)
-	// SetDynamicClient sets the dynamic clientset
-	SetDynamicClient(dynamic.Interface)
-	// SetOpenshiftAppsClient sets the openshift apps clientset
-	SetOpenshiftAppsClient(ocp_clientset.Interface)
-	// SetOpenshiftSecurityClient sets the openshift security clientset
-	SetOpenshiftSecurityClient(ocp_security_clientset.Interface)
-	// SetTalismanClient sets the talisman clientset
-	SetTalismanClient(talismanclientset.Interface)
-	// SetAutopilotClient sets the autopilot clientset
-	SetAutopilotClient(autopilotclientset.Interface)
-	// SetPrometheusClient sets the prometheus clientset
-	SetPrometheusClient(prometheusclient.Interface)
-}
-
 type privateMethods interface {
 	initK8sClient() error
-}
-
-// CustomResource is for creating a Kubernetes TPR/CRD
-type CustomResource struct {
-	// Name of the custom resource
-	Name string
-
-	// ShortNames are short names for the resource.  It must be all lowercase.
-	ShortNames []string
-
-	// Plural of the custom resource in plural
-	Plural string
-
-	// Group the custom resource belongs to
-	Group string
-
-	// Version which should be defined in a const above
-	Version string
-
-	// Scope of the CRD. Namespaced or cluster
-	Scope apiextensionsv1beta1.ResourceScope
-
-	// Kind is the serialized interface of the resource.
-	Kind string
 }
 
 var (
@@ -894,117 +508,6 @@ func NewInstance(config string) (Ops, error) {
 	return newInstance, nil
 }
 
-// SetConfig sets the config and resets the client
-func (k *k8sOps) SetConfig(config *rest.Config) {
-	k.config = config
-	k.client = nil
-}
-
-// SetConfigFromPath takes the path to a kubeconfig file
-// and then internally calls SetConfig to set it
-func (k *k8sOps) SetConfigFromPath(configPath string) error {
-	if configPath == "" {
-		k.SetConfig(nil)
-		return nil
-	}
-	config, err := clientcmd.BuildConfigFromFlags("", configPath)
-	if err != nil {
-		return err
-	}
-
-	k.SetConfig(config)
-	return nil
-}
-
-// SetClient set the k8s clients
-func (k *k8sOps) SetClient(
-	client kubernetes.Interface,
-	snapClient rest.Interface,
-	storkClient storkclientset.Interface,
-	apiExtensionClient apiextensionsclient.Interface,
-	dynamicInterface dynamic.Interface,
-	ocpClient ocp_clientset.Interface,
-	ocpSecurityClient ocp_security_clientset.Interface,
-	autopilotClient autopilotclientset.Interface,
-) {
-	k.client = client
-	k.snapClient = snapClient
-	k.storkClient = storkClient
-	k.apiExtensionClient = apiExtensionClient
-	k.dynamicInterface = dynamicInterface
-	k.ocpClient = ocpClient
-	k.ocpSecurityClient = ocpSecurityClient
-	k.autopilotClient = autopilotClient
-}
-
-// SetBaseClient sets the kubernetes clientset
-func (k *k8sOps) SetBaseClient(client kubernetes.Interface) {
-	k.client = client
-}
-
-// SetSnapshotClient sets the snapshot clientset
-func (k *k8sOps) SetSnapshotClient(snapClient rest.Interface) {
-	k.snapClient = snapClient
-}
-
-// SetStorkClient sets the stork clientset
-func (k *k8sOps) SetStorkClient(storkClient storkclientset.Interface) {
-	k.storkClient = storkClient
-}
-
-// SetApiExtensionsClient sets the api extensions clientset
-func (k *k8sOps) SetApiExtensionsClient(apiExtensionsClient apiextensionsclient.Interface) {
-	k.apiExtensionClient = apiExtensionsClient
-}
-
-// SetDynamicClient sets the dynamic clientset
-func (k *k8sOps) SetDynamicClient(dynamicClient dynamic.Interface) {
-	k.dynamicInterface = dynamicClient
-}
-
-// SetOpenshiftAppsClient sets the openshift apps clientset
-func (k *k8sOps) SetOpenshiftAppsClient(ocpAppsClient ocp_clientset.Interface) {
-	k.ocpClient = ocpAppsClient
-}
-
-// SetOpenshiftSecurityClient sets the openshift security clientset
-func (k *k8sOps) SetOpenshiftSecurityClient(ocpSecurityClient ocp_security_clientset.Interface) {
-	k.ocpSecurityClient = ocpSecurityClient
-}
-
-// SetAutopilotClient sets the autopilot clientset
-func (k *k8sOps) SetAutopilotClient(autopilotClient autopilotclientset.Interface) {
-	k.autopilotClient = autopilotClient
-}
-
-// SetTalismanClient sets the talisman clientset
-func (k *k8sOps) SetTalismanClient(talismanClient talismanclientset.Interface) {
-	k.talismanClient = talismanClient
-}
-
-// SetPrometheusClient sets the prometheus clientset
-func (k *k8sOps) SetPrometheusClient(prometheusClient prometheusclient.Interface) {
-	k.prometheusClient = prometheusClient
-}
-
-// Initialize the k8s client if uninitialized
-func (k *k8sOps) initK8sClient() error {
-	if k.client == nil {
-		err := k.setK8sClient()
-		if err != nil {
-			return err
-		}
-
-		// Quick validation if client connection works
-		_, err = k.client.Discovery().ServerVersion()
-		if err != nil {
-			return fmt.Errorf("failed to connect to k8s server: %s", err)
-		}
-
-	}
-	return nil
-}
-
 func (k *k8sOps) GetVersion() (*version.Info, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
@@ -1012,38 +515,6 @@ func (k *k8sOps) GetVersion() (*version.Info, error) {
 
 	return k.client.Discovery().ServerVersion()
 }
-
-// Security Context Constraints APIs - BEGIN
-
-func (k *k8sOps) getOcpSecurityClient() ocp_securityv1_client.SecurityV1Interface {
-	return k.ocpSecurityClient.SecurityV1()
-}
-
-func (k *k8sOps) ListSecurityContextConstraints() (result *ocp_securityv1_api.SecurityContextConstraintsList, err error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.getOcpSecurityClient().SecurityContextConstraints().List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) GetSecurityContextConstraints(name string) (result *ocp_securityv1_api.SecurityContextConstraints, err error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.getOcpSecurityClient().SecurityContextConstraints().Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) UpdateSecurityContextConstraints(securityContextConstraints *ocp_securityv1_api.SecurityContextConstraints) (result *ocp_securityv1_api.SecurityContextConstraints, err error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.getOcpSecurityClient().SecurityContextConstraints().Update(securityContextConstraints)
-}
-
-// Security Context Constraints APIs - END
 
 // Namespace APIs - BEGIN
 
@@ -1608,23 +1079,23 @@ func (k *k8sOps) PatchService(name, namespace string, jsonPatch []byte) (*v1.Ser
 
 // Deployment APIs - BEGIN
 
-func (k *k8sOps) ListDeployments(namespace string, options meta_v1.ListOptions) (*apps_api.DeploymentList, error) {
+func (k *k8sOps) ListDeployments(namespace string, options meta_v1.ListOptions) (*appsv1.DeploymentList, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().Deployments(namespace).List(options)
+	return k.client.AppsV1().Deployments(namespace).List(options)
 }
 
-func (k *k8sOps) GetDeployment(name, namespace string) (*apps_api.Deployment, error) {
+func (k *k8sOps) GetDeployment(name, namespace string) (*appsv1.Deployment, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().Deployments(namespace).Get(name, meta_v1.GetOptions{})
+	return k.client.AppsV1().Deployments(namespace).Get(name, meta_v1.GetOptions{})
 }
 
-func (k *k8sOps) CreateDeployment(deployment *apps_api.Deployment) (*apps_api.Deployment, error) {
+func (k *k8sOps) CreateDeployment(deployment *appsv1.Deployment) (*appsv1.Deployment, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -1634,7 +1105,7 @@ func (k *k8sOps) CreateDeployment(deployment *apps_api.Deployment) (*apps_api.De
 		ns = v1.NamespaceDefault
 	}
 
-	return k.appsClient().Deployments(ns).Create(deployment)
+	return k.client.AppsV1().Deployments(ns).Create(deployment)
 }
 
 func (k *k8sOps) DeleteDeployment(name, namespace string) error {
@@ -1642,12 +1113,12 @@ func (k *k8sOps) DeleteDeployment(name, namespace string) error {
 		return err
 	}
 
-	return k.appsClient().Deployments(namespace).Delete(name, &meta_v1.DeleteOptions{
+	return k.client.AppsV1().Deployments(namespace).Delete(name, &meta_v1.DeleteOptions{
 		PropagationPolicy: &deleteForegroundPolicy,
 	})
 }
 
-func (k *k8sOps) DescribeDeployment(depName, depNamespace string) (*apps_api.DeploymentStatus, error) {
+func (k *k8sOps) DescribeDeployment(depName, depNamespace string) (*appsv1.DeploymentStatus, error) {
 	dep, err := k.GetDeployment(depName, depNamespace)
 	if err != nil {
 		return nil, err
@@ -1655,14 +1126,14 @@ func (k *k8sOps) DescribeDeployment(depName, depNamespace string) (*apps_api.Dep
 	return &dep.Status, err
 }
 
-func (k *k8sOps) UpdateDeployment(deployment *apps_api.Deployment) (*apps_api.Deployment, error) {
+func (k *k8sOps) UpdateDeployment(deployment *appsv1.Deployment) (*appsv1.Deployment, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
-	return k.appsClient().Deployments(deployment.Namespace).Update(deployment)
+	return k.client.AppsV1().Deployments(deployment.Namespace).Update(deployment)
 }
 
-func (k *k8sOps) ValidateDeployment(deployment *apps_api.Deployment, timeout, retryInterval time.Duration) error {
+func (k *k8sOps) ValidateDeployment(deployment *appsv1.Deployment, timeout, retryInterval time.Duration) error {
 	t := func() (interface{}, bool, error) {
 		dep, err := k.GetDeployment(deployment.Name, deployment.Namespace)
 		if err != nil {
@@ -1763,7 +1234,7 @@ func (k *k8sOps) ValidateDeployment(deployment *apps_api.Deployment, timeout, re
 	return nil
 }
 
-func (k *k8sOps) ValidateTerminatedDeployment(deployment *apps_api.Deployment, timeout, timeBeforeRetry time.Duration) error {
+func (k *k8sOps) ValidateTerminatedDeployment(deployment *appsv1.Deployment, timeout, timeBeforeRetry time.Duration) error {
 	t := func() (interface{}, bool, error) {
 		dep, err := k.GetDeployment(deployment.Name, deployment.Namespace)
 		if err != nil {
@@ -1801,12 +1272,12 @@ func (k *k8sOps) ValidateTerminatedDeployment(deployment *apps_api.Deployment, t
 	return nil
 }
 
-func (k *k8sOps) GetDeploymentPods(deployment *apps_api.Deployment) ([]v1.Pod, error) {
+func (k *k8sOps) GetDeploymentPods(deployment *appsv1.Deployment) ([]v1.Pod, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	rSets, err := k.appsClient().ReplicaSets(deployment.Namespace).List(meta_v1.ListOptions{})
+	rSets, err := k.client.AppsV1().ReplicaSets(deployment.Namespace).List(meta_v1.ListOptions{})
 	if err != nil {
 		return nil, err
 	}
@@ -1822,17 +1293,17 @@ func (k *k8sOps) GetDeploymentPods(deployment *apps_api.Deployment) ([]v1.Pod, e
 	return nil, nil
 }
 
-func (k *k8sOps) GetDeploymentsUsingStorageClass(scName string) ([]apps_api.Deployment, error) {
+func (k *k8sOps) GetDeploymentsUsingStorageClass(scName string) ([]appsv1.Deployment, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	deps, err := k.appsClient().Deployments("").List(meta_v1.ListOptions{})
+	deps, err := k.client.AppsV1().Deployments("").List(meta_v1.ListOptions{})
 	if err != nil {
 		return nil, err
 	}
 
-	var retList []apps_api.Deployment
+	var retList []appsv1.Deployment
 	for _, dep := range deps.Items {
 		for _, v := range dep.Spec.Template.Spec.Volumes {
 			if v.PersistentVolumeClaim == nil {
@@ -1857,273 +1328,22 @@ func (k *k8sOps) GetDeploymentsUsingStorageClass(scName string) ([]apps_api.Depl
 
 // Deployment APIs - END
 
-// DeploymentConfig APIs - BEGIN
-
-func (k *k8sOps) ListDeploymentConfigs(namespace string) (*ocp_appsv1_api.DeploymentConfigList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.ocpAppsClient().DeploymentConfigs(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) GetDeploymentConfig(name, namespace string) (*ocp_appsv1_api.DeploymentConfig, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.ocpAppsClient().DeploymentConfigs(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) CreateDeploymentConfig(deployment *ocp_appsv1_api.DeploymentConfig) (*ocp_appsv1_api.DeploymentConfig, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	ns := deployment.Namespace
-	if len(ns) == 0 {
-		ns = v1.NamespaceDefault
-	}
-
-	return k.ocpAppsClient().DeploymentConfigs(ns).Create(deployment)
-}
-
-func (k *k8sOps) DeleteDeploymentConfig(name, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.ocpAppsClient().DeploymentConfigs(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) DescribeDeploymentConfig(depName, depNamespace string) (*ocp_appsv1_api.DeploymentConfigStatus, error) {
-	dep, err := k.GetDeploymentConfig(depName, depNamespace)
-	if err != nil {
-		return nil, err
-	}
-	return &dep.Status, err
-}
-
-func (k *k8sOps) UpdateDeploymentConfig(deployment *ocp_appsv1_api.DeploymentConfig) (*ocp_appsv1_api.DeploymentConfig, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.ocpAppsClient().DeploymentConfigs(deployment.Namespace).Update(deployment)
-}
-
-func (k *k8sOps) ValidateDeploymentConfig(deployment *ocp_appsv1_api.DeploymentConfig, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		dep, err := k.GetDeploymentConfig(deployment.Name, deployment.Namespace)
-		if err != nil {
-			return "", true, err
-		}
-
-		requiredReplicas := dep.Spec.Replicas
-		shared := false
-
-		if requiredReplicas != 1 {
-			foundPVC := false
-			for _, vol := range dep.Spec.Template.Spec.Volumes {
-				if vol.PersistentVolumeClaim != nil {
-					foundPVC = true
-
-					claim, err := k.client.CoreV1().
-						PersistentVolumeClaims(dep.Namespace).
-						Get(vol.PersistentVolumeClaim.ClaimName, meta_v1.GetOptions{})
-					if err != nil {
-						return "", true, err
-					}
-
-					if k.isPVCShared(claim) {
-						shared = true
-						break
-					}
-				}
-			}
-
-			if foundPVC && !shared {
-				requiredReplicas = 1
-			}
-		}
-
-		pods, err := k.GetDeploymentConfigPods(deployment)
-		if err != nil || pods == nil {
-			return "", true, &ErrAppNotReady{
-				ID:    dep.Name,
-				Cause: fmt.Sprintf("Failed to get pods for deployment. Err: %v", err),
-			}
-		}
-
-		if len(pods) == 0 {
-			return "", true, &ErrAppNotReady{
-				ID:    dep.Name,
-				Cause: "DeploymentConfig has 0 pods",
-			}
-		}
-		podsOverviewString := k.generatePodsOverviewString(pods)
-		if requiredReplicas > dep.Status.AvailableReplicas {
-			return "", true, &ErrAppNotReady{
-				ID: dep.Name,
-				Cause: fmt.Sprintf("Expected replicas: %v Available replicas: %v Current pods overview:\n%s",
-					requiredReplicas, dep.Status.AvailableReplicas, podsOverviewString),
-			}
-		}
-
-		if requiredReplicas > dep.Status.ReadyReplicas {
-			return "", true, &ErrAppNotReady{
-				ID: dep.Name,
-				Cause: fmt.Sprintf("Expected replicas: %v Ready replicas: %v Current pods overview:\n%s",
-					requiredReplicas, dep.Status.ReadyReplicas, podsOverviewString),
-			}
-		}
-
-		if requiredReplicas != dep.Status.UpdatedReplicas && shared {
-			return "", true, &ErrAppNotReady{
-				ID: dep.Name,
-				Cause: fmt.Sprintf("Expected replicas: %v Updated replicas: %v Current pods overview:\n%s",
-					requiredReplicas, dep.Status.UpdatedReplicas, podsOverviewString),
-			}
-		}
-
-		// look for "requiredReplicas" number of pods in ready state
-		var notReadyPods []string
-		var readyCount int32
-		for _, pod := range pods {
-			if !k.IsPodReady(pod) {
-				notReadyPods = append(notReadyPods, pod.Name)
-			} else {
-				readyCount++
-			}
-		}
-
-		if readyCount >= requiredReplicas {
-			return "", false, nil
-		}
-
-		return "", true, &ErrAppNotReady{
-			ID:    dep.Name,
-			Cause: fmt.Sprintf("Pod(s): %#v not yet ready", notReadyPods),
-		}
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (k *k8sOps) ValidateTerminatedDeploymentConfig(deployment *ocp_appsv1_api.DeploymentConfig) error {
-	t := func() (interface{}, bool, error) {
-		dep, err := k.GetDeploymentConfig(deployment.Name, deployment.Namespace)
-		if err != nil {
-			if errors.IsNotFound(err) {
-				return "", false, nil
-			}
-			return "", true, err
-		}
-
-		pods, err := k.GetDeploymentConfigPods(deployment)
-		if err != nil {
-			return "", true, &ErrAppNotTerminated{
-				ID:    dep.Name,
-				Cause: fmt.Sprintf("Failed to get pods for deployment. Err: %v", err),
-			}
-		}
-
-		if pods != nil && len(pods) > 0 {
-			var podNames []string
-			for _, pod := range pods {
-				podNames = append(podNames, pod.Name)
-			}
-			return "", true, &ErrAppNotTerminated{
-				ID:    dep.Name,
-				Cause: fmt.Sprintf("pods: %v are still present", podNames),
-			}
-		}
-
-		return "", false, nil
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, 10*time.Minute, 10*time.Second); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (k *k8sOps) GetDeploymentConfigPods(deployment *ocp_appsv1_api.DeploymentConfig) ([]v1.Pod, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	rSets, err := k.appsClient().ReplicaSets(deployment.Namespace).List(meta_v1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	for _, rSet := range rSets.Items {
-		for _, owner := range rSet.OwnerReferences {
-			if owner.Name == deployment.Name {
-				return k.GetPodsByOwner(rSet.UID, rSet.Namespace)
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-func (k *k8sOps) GetDeploymentConfigsUsingStorageClass(scName string) ([]ocp_appsv1_api.DeploymentConfig, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	deps, err := k.ocpAppsClient().DeploymentConfigs("").List(meta_v1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-
-	var retList []ocp_appsv1_api.DeploymentConfig
-	for _, dep := range deps.Items {
-		for _, v := range dep.Spec.Template.Spec.Volumes {
-			if v.PersistentVolumeClaim == nil {
-				continue
-			}
-
-			pvc, err := k.GetPersistentVolumeClaim(v.PersistentVolumeClaim.ClaimName, dep.Namespace)
-			if err != nil {
-				continue // don't let one bad pvc stop processing
-			}
-
-			sc, err := k.getStorageClassForPVC(pvc)
-			if err == nil && sc.Name == scName {
-				retList = append(retList, dep)
-				break
-			}
-		}
-	}
-
-	return retList, nil
-}
-
-// DeploymentConfig APIs - END
-
 // DaemonSet APIs - BEGIN
 
-func (k *k8sOps) CreateDaemonSet(ds *apps_api.DaemonSet) (*apps_api.DaemonSet, error) {
+func (k *k8sOps) CreateDaemonSet(ds *appsv1.DaemonSet) (*appsv1.DaemonSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().DaemonSets(ds.Namespace).Create(ds)
+	return k.client.AppsV1().DaemonSets(ds.Namespace).Create(ds)
 }
 
-func (k *k8sOps) ListDaemonSets(namespace string, listOpts meta_v1.ListOptions) ([]apps_api.DaemonSet, error) {
+func (k *k8sOps) ListDaemonSets(namespace string, listOpts meta_v1.ListOptions) ([]appsv1.DaemonSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	dsList, err := k.appsClient().DaemonSets(namespace).List(listOpts)
+	dsList, err := k.client.AppsV1().DaemonSets(namespace).List(listOpts)
 	if err != nil {
 		return nil, err
 	}
@@ -2131,7 +1351,7 @@ func (k *k8sOps) ListDaemonSets(namespace string, listOpts meta_v1.ListOptions) 
 	return dsList.Items, nil
 }
 
-func (k *k8sOps) GetDaemonSet(name, namespace string) (*apps_api.DaemonSet, error) {
+func (k *k8sOps) GetDaemonSet(name, namespace string) (*appsv1.DaemonSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -2140,14 +1360,14 @@ func (k *k8sOps) GetDaemonSet(name, namespace string) (*apps_api.DaemonSet, erro
 		namespace = v1.NamespaceDefault
 	}
 
-	ds, err := k.appsClient().DaemonSets(namespace).Get(name, meta_v1.GetOptions{})
+	ds, err := k.client.AppsV1().DaemonSets(namespace).Get(name, meta_v1.GetOptions{})
 	if err != nil {
 		return nil, err
 	}
 	return ds, nil
 }
 
-func (k *k8sOps) GetDaemonSetPods(ds *apps_api.DaemonSet) ([]v1.Pod, error) {
+func (k *k8sOps) GetDaemonSetPods(ds *appsv1.DaemonSet) ([]v1.Pod, error) {
 	return k.GetPodsByOwner(ds.UID, ds.Namespace)
 }
 
@@ -2245,12 +1465,12 @@ func (k *k8sOps) generatePodsOverviewString(pods []v1.Pod) string {
 	return buffer.String()
 }
 
-func (k *k8sOps) UpdateDaemonSet(ds *apps_api.DaemonSet) (*apps_api.DaemonSet, error) {
+func (k *k8sOps) UpdateDaemonSet(ds *appsv1.DaemonSet) (*appsv1.DaemonSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().DaemonSets(ds.Namespace).Update(ds)
+	return k.client.AppsV1().DaemonSets(ds.Namespace).Update(ds)
 }
 
 func (k *k8sOps) DeleteDaemonSet(name, namespace string) error {
@@ -2259,7 +1479,7 @@ func (k *k8sOps) DeleteDaemonSet(name, namespace string) error {
 	}
 
 	policy := meta_v1.DeletePropagationForeground
-	return k.appsClient().DaemonSets(namespace).Delete(
+	return k.client.AppsV1().DaemonSets(namespace).Delete(
 		name,
 		&meta_v1.DeleteOptions{PropagationPolicy: &policy})
 }
@@ -2326,23 +1546,23 @@ func (k *k8sOps) ValidateJob(name, namespace string, timeout time.Duration) erro
 
 // StatefulSet APIs - BEGIN
 
-func (k *k8sOps) ListStatefulSets(namespace string) (*apps_api.StatefulSetList, error) {
+func (k *k8sOps) ListStatefulSets(namespace string) (*appsv1.StatefulSetList, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().StatefulSets(namespace).List(meta_v1.ListOptions{})
+	return k.client.AppsV1().StatefulSets(namespace).List(meta_v1.ListOptions{})
 }
 
-func (k *k8sOps) GetStatefulSet(name, namespace string) (*apps_api.StatefulSet, error) {
+func (k *k8sOps) GetStatefulSet(name, namespace string) (*appsv1.StatefulSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	return k.appsClient().StatefulSets(namespace).Get(name, meta_v1.GetOptions{})
+	return k.client.AppsV1().StatefulSets(namespace).Get(name, meta_v1.GetOptions{})
 }
 
-func (k *k8sOps) CreateStatefulSet(statefulset *apps_api.StatefulSet) (*apps_api.StatefulSet, error) {
+func (k *k8sOps) CreateStatefulSet(statefulset *appsv1.StatefulSet) (*appsv1.StatefulSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -2352,7 +1572,7 @@ func (k *k8sOps) CreateStatefulSet(statefulset *apps_api.StatefulSet) (*apps_api
 		ns = v1.NamespaceDefault
 	}
 
-	return k.appsClient().StatefulSets(ns).Create(statefulset)
+	return k.client.AppsV1().StatefulSets(ns).Create(statefulset)
 }
 
 func (k *k8sOps) DeleteStatefulSet(name, namespace string) error {
@@ -2360,12 +1580,12 @@ func (k *k8sOps) DeleteStatefulSet(name, namespace string) error {
 		return err
 	}
 
-	return k.appsClient().StatefulSets(namespace).Delete(name, &meta_v1.DeleteOptions{
+	return k.client.AppsV1().StatefulSets(namespace).Delete(name, &meta_v1.DeleteOptions{
 		PropagationPolicy: &deleteForegroundPolicy,
 	})
 }
 
-func (k *k8sOps) DescribeStatefulSet(ssetName string, ssetNamespace string) (*apps_api.StatefulSetStatus, error) {
+func (k *k8sOps) DescribeStatefulSet(ssetName string, ssetNamespace string) (*appsv1.StatefulSetStatus, error) {
 	sset, err := k.GetStatefulSet(ssetName, ssetNamespace)
 	if err != nil {
 		return nil, err
@@ -2374,14 +1594,14 @@ func (k *k8sOps) DescribeStatefulSet(ssetName string, ssetNamespace string) (*ap
 	return &sset.Status, err
 }
 
-func (k *k8sOps) UpdateStatefulSet(statefulset *apps_api.StatefulSet) (*apps_api.StatefulSet, error) {
+func (k *k8sOps) UpdateStatefulSet(statefulset *appsv1.StatefulSet) (*appsv1.StatefulSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
-	return k.appsClient().StatefulSets(statefulset.Namespace).Update(statefulset)
+	return k.client.AppsV1().StatefulSets(statefulset.Namespace).Update(statefulset)
 }
 
-func (k *k8sOps) ValidateStatefulSet(statefulset *apps_api.StatefulSet, timeout time.Duration) error {
+func (k *k8sOps) ValidateStatefulSet(statefulset *appsv1.StatefulSet, timeout time.Duration) error {
 	t := func() (interface{}, bool, error) {
 		sset, err := k.GetStatefulSet(statefulset.Name, statefulset.Namespace)
 		if err != nil {
@@ -2440,11 +1660,11 @@ func (k *k8sOps) ValidateStatefulSet(statefulset *apps_api.StatefulSet, timeout 
 	return nil
 }
 
-func (k *k8sOps) GetStatefulSetPods(statefulset *apps_api.StatefulSet) ([]v1.Pod, error) {
+func (k *k8sOps) GetStatefulSetPods(statefulset *appsv1.StatefulSet) ([]v1.Pod, error) {
 	return k.GetPodsByOwner(statefulset.UID, statefulset.Namespace)
 }
 
-func (k *k8sOps) ValidateTerminatedStatefulSet(statefulset *apps_api.StatefulSet, timeout, timeBeforeRetry time.Duration) error {
+func (k *k8sOps) ValidateTerminatedStatefulSet(statefulset *appsv1.StatefulSet, timeout, timeBeforeRetry time.Duration) error {
 	t := func() (interface{}, bool, error) {
 		sset, err := k.GetStatefulSet(statefulset.Name, statefulset.Namespace)
 		if err != nil {
@@ -2483,17 +1703,17 @@ func (k *k8sOps) ValidateTerminatedStatefulSet(statefulset *apps_api.StatefulSet
 	return nil
 }
 
-func (k *k8sOps) GetStatefulSetsUsingStorageClass(scName string) ([]apps_api.StatefulSet, error) {
+func (k *k8sOps) GetStatefulSetsUsingStorageClass(scName string) ([]appsv1.StatefulSet, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
 
-	ss, err := k.appsClient().StatefulSets("").List(meta_v1.ListOptions{})
+	ss, err := k.client.AppsV1().StatefulSets("").List(meta_v1.ListOptions{})
 	if err != nil {
 		return nil, err
 	}
 
-	var retList []apps_api.StatefulSet
+	var retList []appsv1.StatefulSet
 	for _, s := range ss.Items {
 		if s.Spec.VolumeClaimTemplates == nil {
 			continue
@@ -2511,7 +1731,7 @@ func (k *k8sOps) GetStatefulSetsUsingStorageClass(scName string) ([]apps_api.Sta
 	return retList, nil
 }
 
-func (k *k8sOps) GetPVCsForStatefulSet(ss *apps_api.StatefulSet) (*v1.PersistentVolumeClaimList, error) {
+func (k *k8sOps) GetPVCsForStatefulSet(ss *appsv1.StatefulSet) (*v1.PersistentVolumeClaimList, error) {
 	listOptions, err := k.getListOptionsForStatefulSet(ss)
 	if err != nil {
 		return nil, err
@@ -2520,7 +1740,7 @@ func (k *k8sOps) GetPVCsForStatefulSet(ss *apps_api.StatefulSet) (*v1.Persistent
 	return k.getPVCsWithListOptions(ss.Namespace, listOptions)
 }
 
-func (k *k8sOps) ValidatePVCsForStatefulSet(ss *apps_api.StatefulSet, timeout, retryTimeout time.Duration) error {
+func (k *k8sOps) ValidatePVCsForStatefulSet(ss *appsv1.StatefulSet, timeout, retryTimeout time.Duration) error {
 	listOptions, err := k.getListOptionsForStatefulSet(ss)
 	if err != nil {
 		return err
@@ -2551,7 +1771,7 @@ func (k *k8sOps) ValidatePVCsForStatefulSet(ss *apps_api.StatefulSet, timeout, r
 	return nil
 }
 
-func (k *k8sOps) getListOptionsForStatefulSet(ss *apps_api.StatefulSet) (meta_v1.ListOptions, error) {
+func (k *k8sOps) getListOptionsForStatefulSet(ss *appsv1.StatefulSet) (meta_v1.ListOptions, error) {
 	// TODO: Handle MatchExpressions as well
 	labels := ss.Spec.Selector.MatchLabels
 
@@ -3042,7 +2262,7 @@ func (k *k8sOps) WatchPods(namespace string, fn WatchFunc, listOptions meta_v1.L
 
 // StorageClass APIs - BEGIN
 
-func (k *k8sOps) GetStorageClasses(labelSelector map[string]string) (*storage_api.StorageClassList, error) {
+func (k *k8sOps) GetStorageClasses(labelSelector map[string]string) (*storagev1.StorageClassList, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -3052,7 +2272,7 @@ func (k *k8sOps) GetStorageClasses(labelSelector map[string]string) (*storage_ap
 	})
 }
 
-func (k *k8sOps) GetStorageClass(name string) (*storage_api.StorageClass, error) {
+func (k *k8sOps) GetStorageClass(name string) (*storagev1.StorageClass, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -3060,7 +2280,7 @@ func (k *k8sOps) GetStorageClass(name string) (*storage_api.StorageClass, error)
 	return k.client.StorageV1().StorageClasses().Get(name, meta_v1.GetOptions{})
 }
 
-func (k *k8sOps) CreateStorageClass(sc *storage_api.StorageClass) (*storage_api.StorageClass, error) {
+func (k *k8sOps) CreateStorageClass(sc *storagev1.StorageClass) (*storagev1.StorageClass, error) {
 	if err := k.initK8sClient(); err != nil {
 		return nil, err
 	}
@@ -3076,7 +2296,7 @@ func (k *k8sOps) DeleteStorageClass(name string) error {
 	return k.client.StorageV1().StorageClasses().Delete(name, &meta_v1.DeleteOptions{})
 }
 
-func (k *k8sOps) GetStorageClassParams(sc *storage_api.StorageClass) (map[string]string, error) {
+func (k *k8sOps) GetStorageClassParams(sc *storagev1.StorageClass) (map[string]string, error) {
 	sc, err := k.GetStorageClass(sc.Name)
 	if err != nil {
 		return nil, err
@@ -3085,7 +2305,7 @@ func (k *k8sOps) GetStorageClassParams(sc *storage_api.StorageClass) (map[string
 	return sc.Parameters, nil
 }
 
-func (k *k8sOps) ValidateStorageClass(name string) (*storage_api.StorageClass, error) {
+func (k *k8sOps) ValidateStorageClass(name string) (*storagev1.StorageClass, error) {
 	return k.GetStorageClass(name)
 }
 
@@ -3574,362 +2794,7 @@ func (k *k8sOps) DeleteSnapshotData(name string) error {
 		Do().Error()
 }
 
-func (k *k8sOps) GetSnapshotSchedule(name string, namespace string) (*v1alpha1.VolumeSnapshotSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().VolumeSnapshotSchedules(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListSnapshotSchedules(namespace string) (*v1alpha1.VolumeSnapshotScheduleList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().VolumeSnapshotSchedules(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateSnapshotSchedule(snapshotSchedule *v1alpha1.VolumeSnapshotSchedule) (*v1alpha1.VolumeSnapshotSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().VolumeSnapshotSchedules(snapshotSchedule.Namespace).Create(snapshotSchedule)
-}
-
-func (k *k8sOps) UpdateSnapshotSchedule(snapshotSchedule *v1alpha1.VolumeSnapshotSchedule) (*v1alpha1.VolumeSnapshotSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().VolumeSnapshotSchedules(snapshotSchedule.Namespace).Update(snapshotSchedule)
-}
-func (k *k8sOps) DeleteSnapshotSchedule(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().VolumeSnapshotSchedules(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateSnapshotSchedule(name string, namespace string, timeout, retryInterval time.Duration) (
-	map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledVolumeSnapshotStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	t := func() (interface{}, bool, error) {
-		resp, err := k.GetSnapshotSchedule(name, namespace)
-		if err != nil {
-			return nil, true, err
-		}
-
-		if len(resp.Status.Items) == 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("0 snapshots have yet run for the snapshot schedule"),
-				Type:  resp,
-			}
-		}
-
-		failedSnapshots := make([]string, 0)
-		pendingSnapshots := make([]string, 0)
-		for _, snapshotStatuses := range resp.Status.Items {
-			if len(snapshotStatuses) > 0 {
-				status := snapshotStatuses[len(snapshotStatuses)-1]
-				if status == nil {
-					return nil, true, &ErrFailedToValidateCustomSpec{
-						Name:  name,
-						Cause: "SnapshotSchedule has an empty migration in it's most recent status",
-						Type:  resp,
-					}
-				}
-
-				if status.Status == snap_v1.VolumeSnapshotConditionReady {
-					continue
-				}
-
-				if status.Status == snap_v1.VolumeSnapshotConditionError {
-					failedSnapshots = append(failedSnapshots,
-						fmt.Sprintf("snapshot: %s failed. status: %v", status.Name, status.Status))
-				} else {
-					pendingSnapshots = append(pendingSnapshots,
-						fmt.Sprintf("snapshot: %s is not done. status: %v", status.Name, status.Status))
-				}
-			}
-		}
-
-		if len(failedSnapshots) > 0 {
-			return nil, false, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("SnapshotSchedule failed as one or more snapshots have failed. %s",
-					failedSnapshots),
-				Type: resp,
-			}
-		}
-
-		if len(pendingSnapshots) > 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("SnapshotSchedule has certain snapshots pending: %s",
-					pendingSnapshots),
-				Type: resp,
-			}
-		}
-
-		return resp.Status.Items, false, nil
-	}
-
-	ret, err := task.DoRetryWithTimeout(t, timeout, retryInterval)
-	if err != nil {
-		return nil, err
-	}
-
-	snapshots, ok := ret.(map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledVolumeSnapshotStatus)
-	if !ok {
-		return nil, fmt.Errorf("invalid type when checking snapshot schedules: %v", snapshots)
-	}
-
-	return snapshots, nil
-}
-
 // Snapshot APIs - END
-
-// GroupSnapshot APIs - BEGIN
-
-func (k *k8sOps) GetGroupSnapshot(name, namespace string) (*v1alpha1.GroupVolumeSnapshot, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().GroupVolumeSnapshots(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListGroupSnapshots(namespace string) (*v1alpha1.GroupVolumeSnapshotList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().GroupVolumeSnapshots(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateGroupSnapshot(snap *v1alpha1.GroupVolumeSnapshot) (*v1alpha1.GroupVolumeSnapshot, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().GroupVolumeSnapshots(snap.Namespace).Create(snap)
-}
-
-func (k *k8sOps) UpdateGroupSnapshot(snap *v1alpha1.GroupVolumeSnapshot) (*v1alpha1.GroupVolumeSnapshot, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().GroupVolumeSnapshots(snap.Namespace).Update(snap)
-}
-
-func (k *k8sOps) DeleteGroupSnapshot(name, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().GroupVolumeSnapshots(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateGroupSnapshot(name, namespace string, retry bool, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		snap, err := k.GetGroupSnapshot(name, namespace)
-		if err != nil {
-			return "", true, err
-		}
-
-		if len(snap.Status.VolumeSnapshots) == 0 {
-			return "", true, &ErrSnapshotNotReady{
-				ID:    name,
-				Cause: fmt.Sprintf("group snapshot has 0 child snapshots yet"),
-			}
-		}
-
-		if snap.Status.Stage == v1alpha1.GroupSnapshotStageFinal {
-			if snap.Status.Status == v1alpha1.GroupSnapshotSuccessful {
-				// Perform extra check that all child snapshots are also ready
-				notDoneChildSnaps := make([]string, 0)
-				for _, childSnap := range snap.Status.VolumeSnapshots {
-					conditions := childSnap.Conditions
-					if len(conditions) == 0 {
-						notDoneChildSnaps = append(notDoneChildSnaps, childSnap.VolumeSnapshotName)
-						continue
-					}
-
-					lastCondition := conditions[0]
-					if lastCondition.Status != v1.ConditionTrue || lastCondition.Type != snap_v1.VolumeSnapshotConditionReady {
-						notDoneChildSnaps = append(notDoneChildSnaps, childSnap.VolumeSnapshotName)
-						continue
-					}
-				}
-
-				if len(notDoneChildSnaps) > 0 {
-					return "", false, &ErrSnapshotFailed{
-						ID: name,
-						Cause: fmt.Sprintf("group snapshot is marked as successfull "+
-							" but following child volumesnapshots are in pending or error state: %s", notDoneChildSnaps),
-					}
-				}
-
-				return "", false, nil
-			}
-
-			if snap.Status.Status == v1alpha1.GroupSnapshotFailed {
-				return "", false, &ErrSnapshotFailed{
-					ID:    name,
-					Cause: fmt.Sprintf("group snapshot is in failed state"),
-				}
-			}
-		}
-
-		return "", true, &ErrSnapshotNotReady{
-			ID:    name,
-			Cause: fmt.Sprintf("stage: %s status: %s", snap.Status.Stage, snap.Status.Status),
-		}
-	}
-
-	if retry {
-		if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-			return err
-		}
-	} else {
-		if _, _, err := t(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (k *k8sOps) GetSnapshotsForGroupSnapshot(name, namespace string) ([]*snap_v1.VolumeSnapshot, error) {
-	snap, err := k.GetGroupSnapshot(name, namespace)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(snap.Status.VolumeSnapshots) == 0 {
-		return nil, fmt.Errorf("group snapshot: [%s] %s does not have any volume snapshots", namespace, name)
-	}
-
-	snapshots := make([]*snap_v1.VolumeSnapshot, 0)
-	for _, snapStatus := range snap.Status.VolumeSnapshots {
-		snap, err := k.GetSnapshot(snapStatus.VolumeSnapshotName, namespace)
-		if err != nil {
-			return nil, err
-		}
-
-		snapshots = append(snapshots, snap)
-	}
-
-	return snapshots, nil
-}
-
-// GroupSnapshot APIs - END
-
-// Restore Snapshot APIs - BEGIN
-func (k *k8sOps) CreateVolumeSnapshotRestore(snapRestore *v1alpha1.VolumeSnapshotRestore) (*v1alpha1.VolumeSnapshotRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().VolumeSnapshotRestores(snapRestore.Namespace).Create(snapRestore)
-}
-
-func (k *k8sOps) UpdateVolumeSnapshotRestore(snapRestore *v1alpha1.VolumeSnapshotRestore) (*v1alpha1.VolumeSnapshotRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().VolumeSnapshotRestores(snapRestore.Namespace).Update(snapRestore)
-}
-
-func (k *k8sOps) GetVolumeSnapshotRestore(name, namespace string) (*v1alpha1.VolumeSnapshotRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().VolumeSnapshotRestores(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListVolumeSnapshotRestore(namespace string) (*v1alpha1.VolumeSnapshotRestoreList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().VolumeSnapshotRestores(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) DeleteVolumeSnapshotRestore(name, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	return k.storkClient.Stork().VolumeSnapshotRestores(namespace).Delete(name, &meta_v1.DeleteOptions{})
-}
-
-func (k *k8sOps) ValidateVolumeSnapshotRestore(name, namespace string, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		if err := k.initK8sClient(); err != nil {
-			return "", true, err
-		}
-
-		snapRestore, err := k.storkClient.Stork().VolumeSnapshotRestores(namespace).Get(name, meta_v1.GetOptions{})
-		if err != nil {
-			return "", true, err
-		}
-
-		if snapRestore.Status.Status == v1alpha1.VolumeSnapshotRestoreStatusSuccessful {
-			return "", false, nil
-		}
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name: snapRestore.Name,
-			Cause: fmt.Sprintf("VolumeSnapshotRestore failed . Error: %v .Expected status: %v Actual status: %v",
-				err, v1alpha1.VolumeSnapshotRestoreStatusSuccessful, snapRestore.Status.Status),
-			Type: snapRestore,
-		}
-	}
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Restore Snapshot APIs - END
-
-// Rule APIs - BEGIN
-func (k *k8sOps) GetRule(name, namespace string) (*v1alpha1.Rule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, nil
-	}
-
-	return k.storkClient.Stork().Rules(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) CreateRule(rule *v1alpha1.Rule) (*v1alpha1.Rule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, nil
-	}
-
-	return k.storkClient.Stork().Rules(rule.GetNamespace()).Create(rule)
-}
-
-func (k *k8sOps) DeleteRule(name, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return nil
-	}
-
-	return k.storkClient.Stork().Rules(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-// Rule APIs - END
 
 // Secret APIs - BEGIN
 
@@ -4069,335 +2934,6 @@ func (k *k8sOps) WatchConfigMap(configMap *v1.ConfigMap, fn WatchFunc) error {
 
 // ConfigMap APIs - END
 
-// ClusterPair APIs - BEGIN
-func (k *k8sOps) GetClusterPair(name string, namespace string) (*v1alpha1.ClusterPair, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ClusterPairs(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListClusterPairs(namespace string) (*v1alpha1.ClusterPairList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ClusterPairs(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateClusterPair(pair *v1alpha1.ClusterPair) (*v1alpha1.ClusterPair, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ClusterPairs(pair.Namespace).Create(pair)
-}
-
-func (k *k8sOps) UpdateClusterPair(pair *v1alpha1.ClusterPair) (*v1alpha1.ClusterPair, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ClusterPairs(pair.Namespace).Update(pair)
-}
-
-func (k *k8sOps) DeleteClusterPair(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().ClusterPairs(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateClusterPair(name string, namespace string, timeout, retryInterval time.Duration) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	t := func() (interface{}, bool, error) {
-		clusterPair, err := k.GetClusterPair(name, namespace)
-		if err != nil {
-			return "", true, err
-		}
-
-		if clusterPair.Status.SchedulerStatus == v1alpha1.ClusterPairStatusReady &&
-			(clusterPair.Status.StorageStatus == v1alpha1.ClusterPairStatusReady ||
-				clusterPair.Status.StorageStatus == v1alpha1.ClusterPairStatusNotProvided) {
-			return "", false, nil
-		} else if clusterPair.Status.SchedulerStatus == v1alpha1.ClusterPairStatusError ||
-			clusterPair.Status.StorageStatus == v1alpha1.ClusterPairStatusError {
-			return "", true, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("Storage Status: %v \t Scheduler Status: %v", clusterPair.Status.StorageStatus, clusterPair.Status.SchedulerStatus),
-				Type:  clusterPair,
-			}
-		}
-
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  name,
-			Cause: fmt.Sprintf("Storage Status: %v \t Scheduler Status: %v", clusterPair.Status.StorageStatus, clusterPair.Status.SchedulerStatus),
-			Type:  clusterPair,
-		}
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ClusterPair APIs - END
-
-// Migration APIs - BEGIN
-func (k *k8sOps) GetMigration(name string, namespace string) (*v1alpha1.Migration, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().Migrations(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListMigrations(namespace string) (*v1alpha1.MigrationList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().Migrations(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateMigration(migration *v1alpha1.Migration) (*v1alpha1.Migration, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().Migrations(migration.Namespace).Create(migration)
-}
-
-func (k *k8sOps) DeleteMigration(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().Migrations(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) UpdateMigration(migration *v1alpha1.Migration) (*v1alpha1.Migration, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().Migrations(migration.Namespace).Update(migration)
-}
-
-func (k *k8sOps) ValidateMigration(name string, namespace string, timeout, retryInterval time.Duration) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	t := func() (interface{}, bool, error) {
-		resp, err := k.GetMigration(name, namespace)
-		if err != nil {
-			return "", true, err
-		}
-
-		if resp.Status.Status == v1alpha1.MigrationStatusSuccessful {
-			return "", false, nil
-		} else if resp.Status.Status == v1alpha1.MigrationStatusFailed {
-			return "", false, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("Migration Status %v", resp.Status.Status),
-				Type:  resp,
-			}
-		}
-
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  name,
-			Cause: fmt.Sprintf("Migration Status %v", resp.Status.Status),
-			Type:  resp,
-		}
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (k *k8sOps) GetMigrationSchedule(name string, namespace string) (*v1alpha1.MigrationSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().MigrationSchedules(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListMigrationSchedules(namespace string) (*v1alpha1.MigrationScheduleList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().MigrationSchedules(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateMigrationSchedule(migrationSchedule *v1alpha1.MigrationSchedule) (*v1alpha1.MigrationSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().MigrationSchedules(migrationSchedule.Namespace).Create(migrationSchedule)
-}
-
-func (k *k8sOps) UpdateMigrationSchedule(migrationSchedule *v1alpha1.MigrationSchedule) (*v1alpha1.MigrationSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().MigrationSchedules(migrationSchedule.Namespace).Update(migrationSchedule)
-}
-func (k *k8sOps) DeleteMigrationSchedule(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().MigrationSchedules(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateMigrationSchedule(name string, namespace string, timeout, retryInterval time.Duration) (
-	map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledMigrationStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	t := func() (interface{}, bool, error) {
-		resp, err := k.GetMigrationSchedule(name, namespace)
-		if err != nil {
-			return nil, true, err
-		}
-
-		if len(resp.Status.Items) == 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("0 migrations have yet run for the migration schedule"),
-				Type:  resp,
-			}
-		}
-
-		failedMigrations := make([]string, 0)
-		pendingMigrations := make([]string, 0)
-		for _, migrationStatuses := range resp.Status.Items {
-			// The check below assumes that the status will not have a failed migration if the last one succeeded
-			// so just get the last status
-			if len(migrationStatuses) > 0 {
-				status := migrationStatuses[len(migrationStatuses)-1]
-				if status == nil {
-					return nil, true, &ErrFailedToValidateCustomSpec{
-						Name:  name,
-						Cause: "MigrationSchedule has an empty migration in it's most recent status",
-						Type:  resp,
-					}
-				}
-
-				if status.Status == v1alpha1.MigrationStatusSuccessful {
-					continue
-				}
-
-				if status.Status == v1alpha1.MigrationStatusFailed {
-					failedMigrations = append(failedMigrations,
-						fmt.Sprintf("migration: %s failed. status: %v", status.Name, status.Status))
-				} else {
-					pendingMigrations = append(pendingMigrations,
-						fmt.Sprintf("migration: %s is not done. status: %v", status.Name, status.Status))
-				}
-			}
-		}
-
-		if len(failedMigrations) > 0 {
-			return nil, false, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("MigrationSchedule failed as one or more migrations have failed. %s",
-					failedMigrations),
-				Type: resp,
-			}
-		}
-
-		if len(pendingMigrations) > 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("MigrationSchedule has certain migrations pending: %s",
-					pendingMigrations),
-				Type: resp,
-			}
-		}
-
-		return resp.Status.Items, false, nil
-	}
-
-	ret, err := task.DoRetryWithTimeout(t, timeout, retryInterval)
-	if err != nil {
-		return nil, err
-	}
-
-	migrations, ok := ret.(map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledMigrationStatus)
-	if !ok {
-		return nil, fmt.Errorf("invalid type when checking migration schedules: %v", migrations)
-	}
-
-	return migrations, nil
-}
-
-// Migration APIs - END
-
-// SchedulePolicy APIs - BEGIN
-func (k *k8sOps) GetSchedulePolicy(name string) (*v1alpha1.SchedulePolicy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().SchedulePolicies().Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListSchedulePolicies() (*v1alpha1.SchedulePolicyList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().SchedulePolicies().List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateSchedulePolicy(schedulePolicy *v1alpha1.SchedulePolicy) (*v1alpha1.SchedulePolicy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().SchedulePolicies().Create(schedulePolicy)
-}
-
-func (k *k8sOps) DeleteSchedulePolicy(name string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().SchedulePolicies().Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) UpdateSchedulePolicy(schedulePolicy *v1alpha1.SchedulePolicy) (*v1alpha1.SchedulePolicy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().SchedulePolicies().Update(schedulePolicy)
-}
-
-// SchedulePolicy APIs - END
-
 // Event APIs - BEGIN
 // CreateEvent puts an event into k8s etcd
 func (k *k8sOps) CreateEvent(event *v1.Event) (*v1.Event, error) {
@@ -4416,340 +2952,6 @@ func (k *k8sOps) ListEvents(namespace string, opts meta_v1.ListOptions) (*v1.Eve
 }
 
 // Event APIs - END
-
-// CRD APIs - BEGIN
-func (k *k8sOps) CreateCRD(resource CustomResource) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	crdName := fmt.Sprintf("%s.%s", resource.Plural, resource.Group)
-	crd := &apiextensionsv1beta1.CustomResourceDefinition{
-		ObjectMeta: meta_v1.ObjectMeta{
-			Name: crdName,
-		},
-		Spec: apiextensionsv1beta1.CustomResourceDefinitionSpec{
-			Group:   resource.Group,
-			Version: resource.Version,
-			Scope:   resource.Scope,
-			Names: apiextensionsv1beta1.CustomResourceDefinitionNames{
-				Singular:   resource.Name,
-				Plural:     resource.Plural,
-				Kind:       resource.Kind,
-				ShortNames: resource.ShortNames,
-			},
-		},
-	}
-
-	_, err := k.apiExtensionClient.ApiextensionsV1beta1().CustomResourceDefinitions().Create(crd)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (k *k8sOps) RegisterCRD(crd *apiextensionsv1beta1.CustomResourceDefinition) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	_, err := k.apiExtensionClient.ApiextensionsV1beta1().CustomResourceDefinitions().Create(crd)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (k *k8sOps) ValidateCRD(resource CustomResource, timeout, retryInterval time.Duration) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	crdName := fmt.Sprintf("%s.%s", resource.Plural, resource.Group)
-	return wait.Poll(retryInterval, timeout, func() (bool, error) {
-		crd, err := k.apiExtensionClient.ApiextensionsV1beta1().CustomResourceDefinitions().Get(crdName, meta_v1.GetOptions{})
-		if err != nil {
-			return false, err
-		}
-		for _, cond := range crd.Status.Conditions {
-			switch cond.Type {
-			case apiextensionsv1beta1.Established:
-				if cond.Status == apiextensionsv1beta1.ConditionTrue {
-					return true, nil
-				}
-			case apiextensionsv1beta1.NamesAccepted:
-				if cond.Status == apiextensionsv1beta1.ConditionFalse {
-					return false, fmt.Errorf("name conflict: %v", cond.Reason)
-				}
-			}
-		}
-		return false, nil
-	})
-}
-
-func (k *k8sOps) DeleteCRD(fullName string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.apiExtensionClient.ApiextensionsV1beta1().
-		CustomResourceDefinitions().
-		Delete(fullName, &meta_v1.DeleteOptions{PropagationPolicy: &deleteForegroundPolicy})
-}
-
-func (k *k8sOps) CreateVolumePlacementStrategy(spec *talisman_v1beta2.VolumePlacementStrategy) (*talisman_v1beta2.VolumePlacementStrategy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.talismanClient.Portworx().VolumePlacementStrategies().Create(spec)
-}
-
-func (k *k8sOps) UpdateVolumePlacementStrategy(spec *talisman_v1beta2.VolumePlacementStrategy) (*talisman_v1beta2.VolumePlacementStrategy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.talismanClient.Portworx().VolumePlacementStrategies().Update(spec)
-}
-
-func (k *k8sOps) ListVolumePlacementStrategies() (*talisman_v1beta2.VolumePlacementStrategyList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.talismanClient.Portworx().VolumePlacementStrategies().List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) DeleteVolumePlacementStrategy(name string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.talismanClient.Portworx().VolumePlacementStrategies().Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) GetVolumePlacementStrategy(name string) (*talisman_v1beta2.VolumePlacementStrategy, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.talismanClient.Portworx().VolumePlacementStrategies().Get(name, meta_v1.GetOptions{})
-}
-
-// CRD APIs - END
-
-// ClusterDomain CRD - BEGIN
-
-// CreateClusterDomainsStatus creates the ClusterDomainStatus
-func (k *k8sOps) CreateClusterDomainsStatus(clusterDomainsStatus *v1alpha1.ClusterDomainsStatus) (*v1alpha1.ClusterDomainsStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainsStatuses().Create(clusterDomainsStatus)
-}
-
-// GetClusterDomainsStatus gets the ClusterDomainsStatus
-func (k *k8sOps) GetClusterDomainsStatus(name string) (*v1alpha1.ClusterDomainsStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainsStatuses().Get(name, meta_v1.GetOptions{})
-}
-
-// UpdateClusterDomainsStatus updates the ClusterDomainsStatus
-func (k *k8sOps) UpdateClusterDomainsStatus(clusterDomainsStatus *v1alpha1.ClusterDomainsStatus) (*v1alpha1.ClusterDomainsStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainsStatuses().Update(clusterDomainsStatus)
-}
-
-// DeleteClusterDomainsStatus deletes the ClusterDomainsStatus
-func (k *k8sOps) DeleteClusterDomainsStatus(name string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	return k.storkClient.Stork().ClusterDomainsStatuses().Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateClusterDomainsStatus(name string, domainMap map[string]bool, timeout, retryInterval time.Duration) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	t := func() (interface{}, bool, error) {
-		cds, err := k.GetClusterDomainsStatus(name)
-		if err != nil {
-			return "", true, err
-		}
-
-		for _, domainInfo := range cds.Status.ClusterDomainInfos {
-			isActive, _ := domainMap[domainInfo.Name]
-			if isActive {
-				if domainInfo.State != v1alpha1.ClusterDomainActive {
-					return "", true, &ErrFailedToValidateCustomSpec{
-						Name: domainInfo.Name,
-						Cause: fmt.Sprintf("ClusterDomainsStatus mismatch. For domain %v "+
-							"expected to be active found inactive", domainInfo.Name),
-						Type: cds,
-					}
-				}
-			} else {
-				if domainInfo.State != v1alpha1.ClusterDomainInactive {
-					return "", true, &ErrFailedToValidateCustomSpec{
-						Name: domainInfo.Name,
-						Cause: fmt.Sprintf("ClusterDomainsStatus mismatch. For domain %v "+
-							"expected to be inactive found active", domainInfo.Name),
-						Type: cds,
-					}
-				}
-			}
-		}
-
-		return "", false, nil
-
-	}
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-
-	return nil
-
-}
-
-// ListClusterDomainStatuses lists ClusterDomainsStatus
-func (k *k8sOps) ListClusterDomainStatuses() (*v1alpha1.ClusterDomainsStatusList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainsStatuses().List(meta_v1.ListOptions{})
-}
-
-// CreateClusterDomainUpdate creates the ClusterDomainUpdate
-func (k *k8sOps) CreateClusterDomainUpdate(clusterDomainUpdate *v1alpha1.ClusterDomainUpdate) (*v1alpha1.ClusterDomainUpdate, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainUpdates().Create(clusterDomainUpdate)
-}
-
-// GetClusterDomainUpdate gets the ClusterDomainUpdate
-func (k *k8sOps) GetClusterDomainUpdate(name string) (*v1alpha1.ClusterDomainUpdate, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainUpdates().Get(name, meta_v1.GetOptions{})
-}
-
-// UpdateClusterDomainUpdate updates the ClusterDomainUpdate
-func (k *k8sOps) UpdateClusterDomainUpdate(clusterDomainUpdate *v1alpha1.ClusterDomainUpdate) (*v1alpha1.ClusterDomainUpdate, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainUpdates().Update(clusterDomainUpdate)
-}
-
-// DeleteClusterDomainUpdate deletes the ClusterDomainUpdate
-func (k *k8sOps) DeleteClusterDomainUpdate(name string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	return k.storkClient.Stork().ClusterDomainUpdates().Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-// ValidateClusterDomainUpdate validates ClusterDomainUpdate
-func (k *k8sOps) ValidateClusterDomainUpdate(name string, timeout, retryInterval time.Duration) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	t := func() (interface{}, bool, error) {
-		resp, err := k.GetClusterDomainUpdate(name)
-		if err != nil {
-			return "", true, err
-		}
-
-		if resp.Status.Status == v1alpha1.ClusterDomainUpdateStatusSuccessful {
-			return "", false, nil
-		} else if resp.Status.Status == v1alpha1.ClusterDomainUpdateStatusFailed {
-			return "", false, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("ClusterDomainUpdate Status %v", resp.Status.Status),
-				Type:  resp,
-			}
-		}
-
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  name,
-			Cause: fmt.Sprintf("ClusterDomainUpdate Status %v", resp.Status.Status),
-			Type:  resp,
-		}
-	}
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ListClusterDomainUpdates lists ClusterDomainUpdates
-func (k *k8sOps) ListClusterDomainUpdates() (*v1alpha1.ClusterDomainUpdateList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.storkClient.Stork().ClusterDomainUpdates().List(meta_v1.ListOptions{})
-}
-
-// ClusterDomain CRD - END
-
-// AutopilotRule CRD - BEGIN
-
-// CreateAutopilotRule creates the AutopilotRule object
-func (k *k8sOps) CreateAutopilotRule(rule *aut_v1alpaha1.AutopilotRule) (*aut_v1alpaha1.AutopilotRule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.autopilotClient.AutopilotV1alpha1().AutopilotRules().Create(rule)
-}
-
-// GetAutopilotRule gets the AutopilotRule for the provided name
-func (k *k8sOps) GetAutopilotRule(name string) (*aut_v1alpaha1.AutopilotRule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.autopilotClient.AutopilotV1alpha1().AutopilotRules().Get(name, meta_v1.GetOptions{})
-}
-
-// UpdateAutopilotRule updates the AutopilotRule
-func (k *k8sOps) UpdateAutopilotRule(rule *aut_v1alpaha1.AutopilotRule) (*aut_v1alpaha1.AutopilotRule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.autopilotClient.AutopilotV1alpha1().AutopilotRules().Update(rule)
-}
-
-// DeleteAutopilotRule deletes the AutopilotRule of the given name
-func (k *k8sOps) DeleteAutopilotRule(name string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-	return k.autopilotClient.AutopilotV1alpha1().AutopilotRules().Delete(name, &meta_v1.DeleteOptions{})
-}
-
-// ListAutopilotRules lists AutopilotRules
-func (k *k8sOps) ListAutopilotRules() (*aut_v1alpaha1.AutopilotRuleList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	return k.autopilotClient.AutopilotV1alpha1().AutopilotRules().List(meta_v1.ListOptions{})
-}
-
-// AutopilotRule CRD - END
 
 // Object APIs - BEGIN
 
@@ -4804,432 +3006,6 @@ func (k *k8sOps) UpdateObject(object runtime.Object) (runtime.Object, error) {
 
 // Object APIs - END
 
-// BackupLocation APIs - BEGIN
-
-func (k *k8sOps) GetBackupLocation(name string, namespace string) (*v1alpha1.BackupLocation, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	backupLocation, err := k.storkClient.Stork().BackupLocations(namespace).Get(name, meta_v1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	err = backupLocation.UpdateFromSecret(k.client)
-	if err != nil {
-		return nil, err
-	}
-	return backupLocation, nil
-}
-
-func (k *k8sOps) ListBackupLocations(namespace string) (*v1alpha1.BackupLocationList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	backupLocations, err := k.storkClient.Stork().BackupLocations(namespace).List(meta_v1.ListOptions{})
-	if err != nil {
-		return nil, err
-	}
-	for i := range backupLocations.Items {
-		err = backupLocations.Items[i].UpdateFromSecret(k.client)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return backupLocations, nil
-}
-
-func (k *k8sOps) CreateBackupLocation(backupLocation *v1alpha1.BackupLocation) (*v1alpha1.BackupLocation, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().BackupLocations(backupLocation.Namespace).Create(backupLocation)
-}
-
-func (k *k8sOps) DeleteBackupLocation(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().BackupLocations(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) UpdateBackupLocation(backupLocation *v1alpha1.BackupLocation) (*v1alpha1.BackupLocation, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().BackupLocations(backupLocation.Namespace).Update(backupLocation)
-}
-
-func (k *k8sOps) ValidateBackupLocation(name, namespace string, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		if err := k.initK8sClient(); err != nil {
-			return "", true, err
-		}
-
-		resp, err := k.GetBackupLocation(name, namespace)
-		if err != nil {
-			return "", true, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("BackupLocation failed . Error: %v", err),
-				Type:  resp,
-			}
-		}
-		return "", false, nil
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-	return nil
-}
-
-// BackupLocation APIs - END
-
-// ApplicationBackupRestore APIs - BEGIN
-
-func (k *k8sOps) GetApplicationBackup(name string, namespace string) (*v1alpha1.ApplicationBackup, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackups(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListApplicationBackups(namespace string) (*v1alpha1.ApplicationBackupList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackups(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateApplicationBackup(backup *v1alpha1.ApplicationBackup) (*v1alpha1.ApplicationBackup, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackups(backup.Namespace).Create(backup)
-}
-
-func (k *k8sOps) DeleteApplicationBackup(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().ApplicationBackups(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) UpdateApplicationBackup(backup *v1alpha1.ApplicationBackup) (*v1alpha1.ApplicationBackup, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackups(backup.Namespace).Update(backup)
-}
-
-func (k *k8sOps) ValidateApplicationBackup(name, namespace string, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		if err := k.initK8sClient(); err != nil {
-			return "", true, err
-		}
-
-		applicationbackup, err := k.GetApplicationBackup(name, namespace)
-		if err != nil {
-			return "", true, err
-		}
-
-		if applicationbackup.Status.Status == v1alpha1.ApplicationBackupStatusSuccessful {
-			return "", false, nil
-		}
-
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  applicationbackup.Name,
-			Cause: fmt.Sprintf("Application backup failed . Error: %v .Expected status: %v Actual status: %v", err, v1alpha1.ApplicationBackupStatusSuccessful, applicationbackup.Status.Status),
-			Type:  applicationbackup,
-		}
-
-	}
-
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (k *k8sOps) GetApplicationRestore(name string, namespace string) (*v1alpha1.ApplicationRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationRestores(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListApplicationRestores(namespace string) (*v1alpha1.ApplicationRestoreList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationRestores(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateApplicationRestore(restore *v1alpha1.ApplicationRestore) (*v1alpha1.ApplicationRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationRestores(restore.Namespace).Create(restore)
-}
-
-func (k *k8sOps) DeleteApplicationRestore(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().ApplicationRestores(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateApplicationRestore(name, namespace string, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		if err := k.initK8sClient(); err != nil {
-			return "", true, err
-		}
-
-		applicationrestore, err := k.storkClient.Stork().ApplicationRestores(namespace).Get(name, meta_v1.GetOptions{})
-		if err != nil {
-			return "", true, err
-		}
-
-		if applicationrestore.Status.Status == v1alpha1.ApplicationRestoreStatusSuccessful {
-			return "", false, nil
-		}
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  applicationrestore.Name,
-			Cause: fmt.Sprintf("Application restore failed . Error: %v .Expected status: %v Actual status: %v", err, v1alpha1.ApplicationRestoreStatusSuccessful, applicationrestore.Status.Status),
-			Type:  applicationrestore,
-		}
-	}
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (k *k8sOps) UpdateApplicationRestore(restore *v1alpha1.ApplicationRestore) (*v1alpha1.ApplicationRestore, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationRestores(restore.Namespace).Update(restore)
-}
-
-func (k *k8sOps) GetApplicationBackupSchedule(name string, namespace string) (*v1alpha1.ApplicationBackupSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackupSchedules(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListApplicationBackupSchedules(namespace string) (*v1alpha1.ApplicationBackupScheduleList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackupSchedules(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateApplicationBackupSchedule(applicationBackupSchedule *v1alpha1.ApplicationBackupSchedule) (*v1alpha1.ApplicationBackupSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackupSchedules(applicationBackupSchedule.Namespace).Create(applicationBackupSchedule)
-}
-
-func (k *k8sOps) UpdateApplicationBackupSchedule(applicationBackupSchedule *v1alpha1.ApplicationBackupSchedule) (*v1alpha1.ApplicationBackupSchedule, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationBackupSchedules(applicationBackupSchedule.Namespace).Update(applicationBackupSchedule)
-}
-
-func (k *k8sOps) DeleteApplicationBackupSchedule(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().ApplicationBackupSchedules(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) ValidateApplicationBackupSchedule(name string, namespace string, timeout, retryInterval time.Duration) (
-	map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledApplicationBackupStatus, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-	t := func() (interface{}, bool, error) {
-		resp, err := k.GetApplicationBackupSchedule(name, namespace)
-		if err != nil {
-			return nil, true, err
-		}
-
-		if len(resp.Status.Items) == 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name:  name,
-				Cause: fmt.Sprintf("0 backups have yet run for the backup schedule"),
-				Type:  resp,
-			}
-		}
-
-		failedBackups := make([]string, 0)
-		pendingBackups := make([]string, 0)
-		for _, backupStatuses := range resp.Status.Items {
-			// The check below assumes that the status will not have a failed
-			// backup if the last one succeeded so just get the last status
-			if len(backupStatuses) > 0 {
-				status := backupStatuses[len(backupStatuses)-1]
-				if status == nil {
-					return nil, true, &ErrFailedToValidateCustomSpec{
-						Name:  name,
-						Cause: "ApplicationBackupSchedule has an empty backup in it's most recent status",
-						Type:  resp,
-					}
-				}
-
-				if status.Status == v1alpha1.ApplicationBackupStatusSuccessful {
-					continue
-				}
-
-				if status.Status == v1alpha1.ApplicationBackupStatusFailed {
-					failedBackups = append(failedBackups,
-						fmt.Sprintf("backup: %s failed. status: %v", status.Name, status.Status))
-				} else {
-					pendingBackups = append(pendingBackups,
-						fmt.Sprintf("backup: %s is not done. status: %v", status.Name, status.Status))
-				}
-			}
-		}
-
-		if len(failedBackups) > 0 {
-			return nil, false, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("ApplicationBackupSchedule failed as one or more backups have failed. %s",
-					failedBackups),
-				Type: resp,
-			}
-		}
-
-		if len(pendingBackups) > 0 {
-			return nil, true, &ErrFailedToValidateCustomSpec{
-				Name: name,
-				Cause: fmt.Sprintf("ApplicationBackupSchedule has certain migrations pending: %s",
-					pendingBackups),
-				Type: resp,
-			}
-		}
-
-		return resp.Status.Items, false, nil
-	}
-
-	ret, err := task.DoRetryWithTimeout(t, timeout, retryInterval)
-	if err != nil {
-		return nil, err
-	}
-
-	backups, ok := ret.(map[v1alpha1.SchedulePolicyType][]*v1alpha1.ScheduledApplicationBackupStatus)
-	if !ok {
-		return nil, fmt.Errorf("invalid type when checking backup schedules: %v", backups)
-	}
-
-	return backups, nil
-}
-
-// ApplicationBackupRestore APIs - END
-
-// ApplicationClone APIs - BEGIN
-
-func (k *k8sOps) GetApplicationClone(name string, namespace string) (*v1alpha1.ApplicationClone, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationClones(namespace).Get(name, meta_v1.GetOptions{})
-}
-
-func (k *k8sOps) ListApplicationClones(namespace string) (*v1alpha1.ApplicationCloneList, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationClones(namespace).List(meta_v1.ListOptions{})
-}
-
-func (k *k8sOps) CreateApplicationClone(clone *v1alpha1.ApplicationClone) (*v1alpha1.ApplicationClone, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationClones(clone.Namespace).Create(clone)
-}
-
-func (k *k8sOps) DeleteApplicationClone(name string, namespace string) error {
-	if err := k.initK8sClient(); err != nil {
-		return err
-	}
-
-	return k.storkClient.Stork().ApplicationClones(namespace).Delete(name, &meta_v1.DeleteOptions{
-		PropagationPolicy: &deleteForegroundPolicy,
-	})
-}
-
-func (k *k8sOps) UpdateApplicationClone(clone *v1alpha1.ApplicationClone) (*v1alpha1.ApplicationClone, error) {
-	if err := k.initK8sClient(); err != nil {
-		return nil, err
-	}
-
-	return k.storkClient.Stork().ApplicationClones(clone.Namespace).Update(clone)
-}
-
-func (k *k8sOps) ValidateApplicationClone(name, namespace string, timeout, retryInterval time.Duration) error {
-	t := func() (interface{}, bool, error) {
-		if err := k.initK8sClient(); err != nil {
-			return "", true, err
-		}
-
-		applicationclone, err := k.storkClient.Stork().ApplicationClones(namespace).Get(name, meta_v1.GetOptions{})
-		if err != nil {
-			return "", true, err
-		}
-
-		if applicationclone.Status.Status == v1alpha1.ApplicationCloneStatusSuccessful {
-			return "", false, nil
-		}
-		return "", true, &ErrFailedToValidateCustomSpec{
-			Name:  applicationclone.Name,
-			Cause: fmt.Sprintf("Application Clone failed . Error: %v .Expected status: %v Actual status: %v", err, v1alpha1.ApplicationCloneStatusSuccessful, applicationclone.Status.Status),
-			Type:  applicationclone,
-		}
-	}
-	if _, err := task.DoRetryWithTimeout(t, timeout, retryInterval); err != nil {
-		return err
-	}
-	return nil
-}
-
-// ApplicationClone APIs - END
-
 // VolumeAttachment APIs - START
 
 func (k *k8sOps) ListVolumeAttachments() (*storagev1beta1.VolumeAttachmentList, error) {
@@ -5274,158 +3050,6 @@ func (k *k8sOps) UpdateVolumeAttachmentStatus(volumeAttachment *storagev1beta1.V
 
 // VolumeAttachment APIs - END
 
-func (k *k8sOps) appsClient() appsv1_client.AppsV1Interface {
-	return k.client.AppsV1()
-}
-
-func (k *k8sOps) ocpAppsClient() ocp_appsv1_client.AppsV1Interface {
-	return k.ocpClient.AppsV1()
-}
-
-// setK8sClient instantiates a k8s client
-func (k *k8sOps) setK8sClient() error {
-	var err error
-
-	if k.config != nil {
-		err = k.loadClientFor(k.config)
-	} else {
-		kubeconfig := os.Getenv("KUBECONFIG")
-		if len(kubeconfig) > 0 {
-			err = k.loadClientFromKubeconfig(kubeconfig)
-		} else {
-			err = k.loadClientFromServiceAccount()
-		}
-
-	}
-	if err != nil {
-		return err
-	}
-
-	if k.client == nil {
-		return ErrK8SApiAccountNotSet
-	}
-
-	return nil
-}
-
-// loadClientFromServiceAccount loads a k8s client from a ServiceAccount specified in the pod running px
-func (k *k8sOps) loadClientFromServiceAccount() error {
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		return err
-	}
-
-	k.config = config
-	return k.loadClientFor(config)
-}
-
-func (k *k8sOps) loadClientFromKubeconfig(kubeconfig string) error {
-	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
-	if err != nil {
-		return err
-	}
-
-	k.config = config
-	return k.loadClientFor(config)
-}
-
-func (k *k8sOps) loadClientFor(config *rest.Config) error {
-	var err error
-	k.client, err = kubernetes.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.snapClient, _, err = snap_client.NewClient(config)
-	if err != nil {
-		return err
-	}
-
-	k.storkClient, err = storkclientset.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.talismanClient, err = talismanclientset.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.apiExtensionClient, err = apiextensionsclient.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.dynamicInterface, err = dynamic.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.ocpClient, err = ocp_clientset.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.ocpSecurityClient, err = ocp_security_clientset.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.autopilotClient, err = autopilotclientset.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	k.prometheusClient, err = prometheusclient.NewForConfig(config)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func roundUpSize(volumeSizeBytes int64, allocationUnitBytes int64) int64 {
-	return (volumeSizeBytes + allocationUnitBytes - 1) / allocationUnitBytes
-}
-
-// getLocalIPList returns the list of local IP addresses, and optionally includes local hostname.
-func getLocalIPList(includeHostname bool) ([]string, error) {
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		return nil, err
-	}
-	ipList := make([]string, 0, len(ifaces))
-	for _, i := range ifaces {
-		addrs, err := i.Addrs()
-		if err != nil {
-			logrus.WithError(err).Warnf("Error listing address for %s (cont.)", i.Name)
-			continue
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-			// process IP address
-			if ip != nil && !ip.IsLoopback() && !ip.IsUnspecified() {
-				ipList = append(ipList, ip.String())
-			}
-		}
-	}
-
-	if includeHostname {
-		hn, err := os.Hostname()
-		if err == nil && hn != "" && !strings.HasPrefix(hn, "localhost") {
-			ipList = append(ipList, hn)
-		}
-	}
-
-	return ipList, nil
-}
-
 // isAnyVolumeUsingVolumePlugin returns true if any of the given volumes is using a storage class for the given plugin
 //	In case errors are found while looking up a particular volume, the function ignores the errors as the goal is to
 //	find if there is any match or not
@@ -5447,7 +3071,7 @@ func (k *k8sOps) isAnyVolumeUsingVolumePlugin(volumes []v1.Volume, volumeNamespa
 	return false
 }
 
-func (k *k8sOps) getStorageClassForPVC(pvc *v1.PersistentVolumeClaim) (*storage_api.StorageClass, error) {
+func (k *k8sOps) getStorageClassForPVC(pvc *v1.PersistentVolumeClaim) (*storagev1.StorageClass, error) {
 	var scName string
 	if pvc.Spec.StorageClassName != nil && len(*pvc.Spec.StorageClassName) > 0 {
 		scName = *pvc.Spec.StorageClassName
@@ -5460,13 +3084,4 @@ func (k *k8sOps) getStorageClassForPVC(pvc *v1.PersistentVolumeClaim) (*storage_
 	}
 
 	return k.GetStorageClass(scName)
-}
-
-func mapToCSV(in map[string]string) string {
-	var items []string
-	for k, v := range in {
-		items = append(items, fmt.Sprintf("%s=%s", k, v))
-	}
-
-	return strings.Join(items, ",")
 }
