+++ k8s/util.go
@@ -1,0 +1,61 @@
+package k8s
+
+import (
+	"fmt"
+	"net"
+	"os"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+)
+
+func mapToCSV(in map[string]string) string {
+	var items []string
+	for k, v := range in {
+		items = append(items, fmt.Sprintf("%s=%s", k, v))
+	}
+
+	return strings.Join(items, ",")
+}
+
+func roundUpSize(volumeSizeBytes int64, allocationUnitBytes int64) int64 {
+	return (volumeSizeBytes + allocationUnitBytes - 1) / allocationUnitBytes
+}
+
+// getLocalIPList returns the list of local IP addresses, and optionally includes local hostname.
+func getLocalIPList(includeHostname bool) ([]string, error) {
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		return nil, err
+	}
+	ipList := make([]string, 0, len(ifaces))
+	for _, i := range ifaces {
+		addrs, err := i.Addrs()
+		if err != nil {
+			logrus.WithError(err).Warnf("Error listing address for %s (cont.)", i.Name)
+			continue
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			// process IP address
+			if ip != nil && !ip.IsLoopback() && !ip.IsUnspecified() {
+				ipList = append(ipList, ip.String())
+			}
+		}
+	}
+
+	if includeHostname {
+		hn, err := os.Hostname()
+		if err == nil && hn != "" && !strings.HasPrefix(hn, "localhost") {
+			ipList = append(ipList, hn)
+		}
+	}
+
+	return ipList, nil
+}
